test: organize Realtime session tests by responsibility
