+++ tests/realtime/test_session.py
@@ -1,21 +1,17 @@
+"""Session lifecycle, model settings, event forwarding, and tool dispatch tests."""
+
 import asyncio
-import dataclasses
 import json
-import logging
-import threading
 import traceback
 from pathlib import Path
 from typing import Any, Literal, cast
-from unittest.mock import AsyncMock, Mock, PropertyMock, patch
+from unittest.mock import AsyncMock, Mock, patch
 
 import pytest
-from pydantic import BaseModel, ConfigDict
 
 import agents._debug as _debug
-from agents._tool_identity import get_function_tool_lookup_key_for_tool
 from agents.agent import AgentBase
 from agents.exceptions import ModelBehaviorError, ToolTimeoutError, UserError
-from agents.guardrail import GuardrailFunctionOutput, OutputGuardrail
 from agents.handoffs import Handoff
 from agents.realtime import realtime_handoff
 from agents.realtime.agent import RealtimeAgent
@@ -27,23 +23,9 @@ from agents.realtime.events import (
     RealtimeAudioEnd,
     RealtimeAudioInterrupted,
     RealtimeError,
-    RealtimeGuardrailTripped,
-    RealtimeHistoryAdded,
-    RealtimeHistoryUpdated,
     RealtimeRawModelEvent,
-    RealtimeToolApprovalRequired,
     RealtimeToolEnd,
     RealtimeToolStart,
-)
-from agents.realtime.items import (
-    AssistantAudio,
-    AssistantMessageItem,
-    AssistantText,
-    InputAudio,
-    InputText,
-    RealtimeItem,
-    RealtimeToolCallItem,
-    UserMessageItem,
 )
 from agents.realtime.model import RealtimeModel, RealtimeModelConfig
 from agents.realtime.model_events import (
@@ -54,10 +36,7 @@ from agents.realtime.model_events import (
     RealtimeModelEndOfStreamEvent,
     RealtimeModelErrorEvent,
     RealtimeModelInputAudioTranscriptionCompletedEvent,
-    RealtimeModelItemDeletedEvent,
-    RealtimeModelItemUpdatedEvent,
     RealtimeModelOtherEvent,
-    RealtimeModelOutputTextDeltaEvent,
     RealtimeModelToolCallEvent,
     RealtimeModelTranscriptDeltaEvent,
     RealtimeModelTurnEndedEvent,
@@ -72,12 +51,9 @@ from agents.realtime.model_inputs import (
     RealtimeModelSendUserInput,
 )
 from agents.realtime.session import (
-    REJECTION_MESSAGE,
     RealtimeSession,
     _PendingToolOutputSendError,
-    _serialize_tool_output,
 )
-from agents.realtime.testing import RealtimeConnectCall, ScriptedRealtimeModel
 from agents.run_context import RunContextWrapper
 from agents.tool import FunctionTool, function_tool, tool_namespace
 from agents.tool_context import ToolContext
@@ -88,18 +64,18 @@ from agents.tool_guardrails import (
 )
 from agents.usage import Usage
 
+from . import session_test_support
+from .session_test_support import (
+    _DummyModel,
+    _named_function_tool,
+    _sent_tool_output_strings,
+    _set_default_timeout_fields,
+)
 
-class _DummyModel(ScriptedRealtimeModel):
-    def __init__(self) -> None:
-        super().__init__(strict=False)
-
-    @property
-    def events(self) -> tuple[Any, ...]:
-        return self.sent_events
-
-    @property
-    def connect_options(self) -> RealtimeConnectCall | None:
-        return self.connect_calls[-1] if self.connect_calls else None
+# Bind shared fixtures explicitly so unrelated Realtime modules do not inherit them.
+mock_agent = session_test_support.mock_agent
+mock_function_tool = session_test_support.mock_function_tool
+mock_model = session_test_support.mock_model
 
 
 class _FailingConnectModel(_DummyModel):
@@ -613,51 +589,6 @@ async def test_model_event_cannot_mutate_history_after_raw_event_enqueue_and_clo
 
     assert session._closed
     assert session._history == []
-
-
-@pytest.mark.asyncio
-async def test_transcription_completed_adds_new_user_item():
-    model = _DummyModel()
-    agent = RealtimeAgent(name="agent")
-    session = RealtimeSession(model, agent, None)
-
-    event = RealtimeModelInputAudioTranscriptionCompletedEvent(item_id="item1", transcript="hello")
-    await session.on_event(event)
-
-    # Should have appended a new user item
-    assert len(session._history) == 1
-    assert session._history[0].type == "message"
-    assert session._history[0].role == "user"
-
-
-class _FakeAudio:
-    # Looks like an audio part but is not an InputAudio/AssistantAudio instance
-    type = "audio"
-    transcript = None
-
-
-@pytest.mark.asyncio
-async def test_item_updated_merge_exception_path_logs_error(monkeypatch):
-    monkeypatch.setattr(_debug, "DONT_LOG_MODEL_DATA", True)
-    model = _DummyModel()
-    agent = RealtimeAgent(name="agent")
-    session = RealtimeSession(model, agent, None)
-
-    # existing assistant message with transcript to preserve
-    existing = AssistantMessageItem(
-        item_id="a1", role="assistant", content=[AssistantAudio(audio=None, transcript="t")]
-    )
-    session._history = [existing]
-
-    # incoming message with a deliberately bogus content entry to trigger assertion path
-    incoming = AssistantMessageItem(
-        item_id="a1", role="assistant", content=[AssistantAudio(audio=None, transcript=None)]
-    )
-    incoming.content[0] = cast(Any, _FakeAudio())
-
-    with patch("agents.realtime.session.logger") as mock_logger:
-        await session.on_event(RealtimeModelItemUpdatedEvent(item=incoming))
-        mock_logger.error.assert_called_once_with("%s", "Error merging transcripts", stacklevel=3)
 
 
 @pytest.mark.asyncio
@@ -1290,95 +1221,6 @@ async def test_aenter_removes_listener_when_connect_fails(exc: BaseException):
     assert model.listeners == ()
 
 
-class RecordingRealtimeModel(ScriptedRealtimeModel):
-    def __init__(self):
-        super().__init__(strict=False)
-        # Legacy tracking for tests that haven't been updated yet
-        self.sent_messages = []
-        self.sent_audio = []
-        self.sent_tool_outputs = []
-        self.interrupts_called = 0
-        self.retired_audio_response_ids = []
-
-    async def send_event(self, event):
-        from agents.realtime.model_inputs import (
-            RealtimeModelSendAudio,
-            RealtimeModelSendInterrupt,
-            RealtimeModelSendToolOutput,
-            RealtimeModelSendUserInput,
-        )
-
-        self._sent_events.append(self._snapshot_send_event(event))
-
-        # Update legacy tracking for compatibility
-        if isinstance(event, RealtimeModelSendUserInput):
-            self.sent_messages.append(event.user_input)
-        elif isinstance(event, RealtimeModelSendAudio):
-            self.sent_audio.append((event.audio, event.commit))
-        elif isinstance(event, RealtimeModelSendToolOutput):
-            self.sent_tool_outputs.append((event.tool_call, event.output, event.start_response))
-        elif isinstance(event, RealtimeModelSendInterrupt):
-            self.interrupts_called += 1
-
-    async def send_event_if(self, event, send_if):
-        if not send_if():
-            return False
-        await self.send_event(event)
-        return True
-
-    def _retire_response_audio(self, response_id: str) -> None:
-        self.retired_audio_response_ids.append(response_id)
-
-
-@pytest.fixture
-def mock_agent():
-    agent = Mock(spec=RealtimeAgent)
-    agent.get_all_tools = AsyncMock(return_value=[])
-
-    type(agent).handoffs = PropertyMock(return_value=[])
-    type(agent).output_guardrails = PropertyMock(return_value=[])
-    return agent
-
-
-@pytest.fixture
-def mock_model():
-    return RecordingRealtimeModel()
-
-
-def _set_default_timeout_fields(tool: Mock) -> Mock:
-    tool.timeout_seconds = None
-    tool.timeout_behavior = "error_as_result"
-    tool.timeout_error_function = None
-    return tool
-
-
-def _named_function_tool(
-    name: str,
-    output: str,
-    *,
-    needs_approval: bool = False,
-) -> FunctionTool:
-    def tool_func() -> str:
-        return output
-
-    tool = function_tool(tool_func, name_override=name)
-    tool.needs_approval = needs_approval
-    return tool
-
-
-def _sent_tool_output_strings(model: RecordingRealtimeModel) -> list[str]:
-    return [output for _call, output, _start_response in model.sent_tool_outputs]
-
-
-@pytest.fixture
-def mock_function_tool():
-    tool = _set_default_timeout_fields(Mock(spec=FunctionTool))
-    tool.name = "test_function"
-    tool.on_invoke_tool = AsyncMock(return_value="function_result")
-    tool.needs_approval = False
-    return tool
-
-
 @pytest.fixture
 def mock_handoff():
     handoff = Mock(spec=Handoff)
@@ -1513,178 +1355,6 @@ class TestEventHandling:
         assert len(first_raw.info.context.usage.request_usage_entries) == 2
 
     @pytest.mark.asyncio
-    async def test_transcription_completed_event_updates_history(self, mock_model, mock_agent):
-        """Test that transcription completed events update history and emit events"""
-        session = RealtimeSession(
-            mock_model, mock_agent, None, run_config={"async_tool_calls": False}
-        )
-
-        # Set up initial history with an audio message
-        initial_item = UserMessageItem(
-            item_id="item_1", role="user", content=[InputAudio(transcript=None)]
-        )
-        session._history = [initial_item]
-
-        # Create transcription completed event
-        transcription_event = RealtimeModelInputAudioTranscriptionCompletedEvent(
-            item_id="item_1", transcript="Hello world"
-        )
-
-        await session.on_event(transcription_event)
-
-        # Check that history was updated
-        assert len(session._history) == 1
-        updated_item = session._history[0]
-        assert updated_item.content[0].transcript == "Hello world"  # type: ignore
-        assert updated_item.status == "completed"  # type: ignore
-
-        # Should have 2 events: raw + history updated
-        assert session._event_queue.qsize() == 2
-
-        await session._event_queue.get()  # raw event
-        history_event = await session._event_queue.get()
-        assert isinstance(history_event, RealtimeHistoryUpdated)
-        assert len(history_event.history) == 1
-
-    @pytest.mark.asyncio
-    async def test_item_updated_event_adds_new_item(self, mock_model, mock_agent):
-        """Test that item_updated events add new items to history"""
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={"async_tool_calls": False},
-        )
-
-        new_item = AssistantMessageItem(
-            item_id="new_item", role="assistant", content=[AssistantText(text="Hello")]
-        )
-
-        item_updated_event = RealtimeModelItemUpdatedEvent(item=new_item)
-
-        await session.on_event(item_updated_event)
-
-        # Check that item was added to history
-        assert len(session._history) == 1
-        assert session._history[0] == new_item
-
-        # Should have 2 events: raw + history added
-        assert session._event_queue.qsize() == 2
-
-        await session._event_queue.get()  # raw event
-        history_event = await session._event_queue.get()
-        assert isinstance(history_event, RealtimeHistoryAdded)
-        assert history_event.item == new_item
-
-    @pytest.mark.asyncio
-    async def test_item_updated_event_updates_existing_item(self, mock_model, mock_agent):
-        """Test that item_updated events update existing items in history"""
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={"async_tool_calls": False},
-        )
-
-        # Set up initial history
-        initial_item = AssistantMessageItem(
-            item_id="existing_item", role="assistant", content=[AssistantText(text="Initial")]
-        )
-        session._history = [initial_item]
-
-        # Create updated version
-        updated_item = AssistantMessageItem(
-            item_id="existing_item", role="assistant", content=[AssistantText(text="Updated")]
-        )
-
-        item_updated_event = RealtimeModelItemUpdatedEvent(item=updated_item)
-
-        await session.on_event(item_updated_event)
-
-        # Check that item was updated
-        assert len(session._history) == 1
-        updated_item = cast(AssistantMessageItem, session._history[0])
-        assert updated_item.content[0].text == "Updated"  # type: ignore
-
-        # Should have 2 events: raw + history updated (not added)
-        assert session._event_queue.qsize() == 2
-
-        await session._event_queue.get()  # raw event
-        history_event = await session._event_queue.get()
-        assert isinstance(history_event, RealtimeHistoryUpdated)
-
-    @pytest.mark.asyncio
-    async def test_item_updated_event_completes_tool_call(self, mock_model, mock_agent):
-        """The transport reuses one item for a tool call and its output, so the second
-        item_updated must land in history."""
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={"async_tool_calls": False},
-        )
-
-        in_progress = RealtimeToolCallItem(
-            item_id="fc_1",
-            previous_item_id=None,
-            call_id="call_1",
-            type="function_call",
-            status="in_progress",
-            arguments='{"city": "Oakland"}',
-            name="get_weather",
-            output=None,
-        )
-        await session.on_event(RealtimeModelItemUpdatedEvent(item=in_progress))
-
-        completed = in_progress.model_copy(update={"status": "completed", "output": "sunny"})
-        await session.on_event(RealtimeModelItemUpdatedEvent(item=completed))
-
-        assert len(session._history) == 1
-        stored = cast(RealtimeToolCallItem, session._history[0])
-        assert stored.status == "completed"
-        assert stored.output == "sunny"
-
-        # raw + history added, then raw + history updated.
-        assert session._event_queue.qsize() == 4
-        await session._event_queue.get()  # raw event
-        assert isinstance(await session._event_queue.get(), RealtimeHistoryAdded)
-        await session._event_queue.get()  # raw event
-        history_event = await session._event_queue.get()
-        assert isinstance(history_event, RealtimeHistoryUpdated)
-        assert cast(RealtimeToolCallItem, history_event.history[0]).output == "sunny"
-
-    @pytest.mark.asyncio
-    async def test_item_deleted_event_removes_item(self, mock_model, mock_agent):
-        """Test that item_deleted events remove items from history"""
-        session = RealtimeSession(mock_model, mock_agent, None)
-
-        # Set up initial history with multiple items
-        item1 = AssistantMessageItem(
-            item_id="item_1", role="assistant", content=[AssistantText(text="First")]
-        )
-        item2 = AssistantMessageItem(
-            item_id="item_2", role="assistant", content=[AssistantText(text="Second")]
-        )
-        session._history = [item1, item2]
-
-        # Delete first item
-        delete_event = RealtimeModelItemDeletedEvent(item_id="item_1")
-
-        await session.on_event(delete_event)
-
-        # Check that item was removed
-        assert len(session._history) == 1
-        assert session._history[0].item_id == "item_2"
-
-        # Should have 2 events: raw + history updated
-        assert session._event_queue.qsize() == 2
-
-        await session._event_queue.get()  # raw event
-        history_event = await session._event_queue.get()
-        assert isinstance(history_event, RealtimeHistoryUpdated)
-        assert len(history_event.history) == 1
-
-    @pytest.mark.asyncio
     async def test_ignored_events_only_generate_raw_events(self, mock_model, mock_agent):
         """Test that ignored events (transcript_delta, connection_status, other) only generate raw
         events"""
@@ -1813,381 +1483,6 @@ class TestEventHandling:
         raw_event = await session._event_queue.get()
         assert isinstance(raw_event, RealtimeRawModelEvent)
         assert raw_event.data == function_call_event
-
-
-class TestHistoryManagement:
-    """Test suite for history management and audio transcription in
-    RealtimeSession._get_new_history"""
-
-    def test_merge_transcript_into_existing_audio_message(self):
-        """Test merging audio transcript into existing placeholder input_audio message"""
-        # Create initial history with audio message without transcript
-        initial_item = UserMessageItem(
-            item_id="item_1",
-            role="user",
-            content=[
-                InputText(text="Before audio"),
-                InputAudio(transcript=None, audio="audio_data"),
-                InputText(text="After audio"),
-            ],
-        )
-        old_history = [initial_item]
-
-        # Create transcription completed event
-        transcription_event = RealtimeModelInputAudioTranscriptionCompletedEvent(
-            item_id="item_1", transcript="Hello world"
-        )
-
-        # Apply the history update
-        new_history = RealtimeSession._get_new_history(
-            cast(list[RealtimeItem], old_history), transcription_event
-        )
-
-        # Verify the transcript was merged
-        assert len(new_history) == 1
-        updated_item = cast(UserMessageItem, new_history[0])
-        assert updated_item.item_id == "item_1"
-        assert hasattr(updated_item, "status") and updated_item.status == "completed"
-        assert len(updated_item.content) == 3
-
-        # Check that audio content got transcript but other content unchanged
-        assert cast(InputText, updated_item.content[0]).text == "Before audio"
-        assert cast(InputAudio, updated_item.content[1]).transcript == "Hello world"
-        # Should preserve audio data
-        assert cast(InputAudio, updated_item.content[1]).audio == "audio_data"
-        assert cast(InputText, updated_item.content[2]).text == "After audio"
-
-    def test_merge_transcript_preserves_other_items(self):
-        """Test that merging transcript preserves other items in history"""
-        # Create history with multiple items
-        item1 = UserMessageItem(
-            item_id="item_1", role="user", content=[InputText(text="First message")]
-        )
-        item2 = UserMessageItem(
-            item_id="item_2", role="user", content=[InputAudio(transcript=None)]
-        )
-        item3 = AssistantMessageItem(
-            item_id="item_3", role="assistant", content=[AssistantText(text="Third message")]
-        )
-        old_history = [item1, item2, item3]
-
-        # Create transcription event for item_2
-        transcription_event = RealtimeModelInputAudioTranscriptionCompletedEvent(
-            item_id="item_2", transcript="Transcribed audio"
-        )
-
-        new_history = RealtimeSession._get_new_history(
-            cast(list[RealtimeItem], old_history), transcription_event
-        )
-
-        # Should have same number of items
-        assert len(new_history) == 3
-
-        # First and third items should be unchanged
-        assert new_history[0] == item1
-        assert new_history[2] == item3
-
-        # Second item should have transcript
-        updated_item2 = cast(UserMessageItem, new_history[1])
-        assert updated_item2.item_id == "item_2"
-        assert cast(InputAudio, updated_item2.content[0]).transcript == "Transcribed audio"
-        assert hasattr(updated_item2, "status") and updated_item2.status == "completed"
-
-    def test_merge_transcript_only_affects_matching_audio_content(self):
-        """Test that transcript merge only affects audio content, not text content"""
-        # Create item with mixed content including multiple audio items
-        item = UserMessageItem(
-            item_id="item_1",
-            role="user",
-            content=[
-                InputText(text="Text content"),
-                InputAudio(transcript=None, audio="audio1"),
-                InputAudio(transcript="existing", audio="audio2"),
-                InputText(text="More text"),
-            ],
-        )
-        old_history = [item]
-
-        transcription_event = RealtimeModelInputAudioTranscriptionCompletedEvent(
-            item_id="item_1", transcript="New transcript"
-        )
-
-        new_history = RealtimeSession._get_new_history(
-            cast(list[RealtimeItem], old_history), transcription_event
-        )
-
-        updated_item = cast(UserMessageItem, new_history[0])
-
-        # Text content should be unchanged
-        assert cast(InputText, updated_item.content[0]).text == "Text content"
-        assert cast(InputText, updated_item.content[3]).text == "More text"
-
-        # All audio content should have the new transcript (current implementation overwrites all)
-        assert cast(InputAudio, updated_item.content[1]).transcript == "New transcript"
-        assert (
-            cast(InputAudio, updated_item.content[2]).transcript == "New transcript"
-        )  # Implementation overwrites existing
-
-    def test_update_existing_item_by_id(self):
-        """Test updating an existing item by item_id"""
-        # Create initial history
-        original_item = AssistantMessageItem(
-            item_id="item_1", role="assistant", content=[AssistantText(text="Original")]
-        )
-        old_history = [original_item]
-
-        # Create updated version of same item
-        updated_item = AssistantMessageItem(
-            item_id="item_1", role="assistant", content=[AssistantText(text="Updated")]
-        )
-
-        new_history = RealtimeSession._get_new_history(
-            cast(list[RealtimeItem], old_history), updated_item
-        )
-
-        # Should have same number of items
-        assert len(new_history) == 1
-
-        # Item should be updated
-        result_item = cast(AssistantMessageItem, new_history[0])
-        assert result_item.item_id == "item_1"
-        assert result_item.content[0].text == "Updated"  # type: ignore
-
-    def test_update_existing_item_preserves_order(self):
-        """Test that updating existing item preserves its position in history"""
-        # Create history with multiple items
-        item1 = AssistantMessageItem(
-            item_id="item_1", role="assistant", content=[AssistantText(text="First")]
-        )
-        item2 = AssistantMessageItem(
-            item_id="item_2", role="assistant", content=[AssistantText(text="Second")]
-        )
-        item3 = AssistantMessageItem(
-            item_id="item_3", role="assistant", content=[AssistantText(text="Third")]
-        )
-        old_history = [item1, item2, item3]
-
-        # Update middle item
-        updated_item2 = AssistantMessageItem(
-            item_id="item_2", role="assistant", content=[AssistantText(text="Updated Second")]
-        )
-
-        new_history = RealtimeSession._get_new_history(
-            cast(list[RealtimeItem], old_history), updated_item2
-        )
-
-        # Should have same number of items in same order
-        assert len(new_history) == 3
-        assert new_history[0].item_id == "item_1"
-        assert new_history[1].item_id == "item_2"
-        assert new_history[2].item_id == "item_3"
-
-        # Middle item should be updated
-        updated_result = cast(AssistantMessageItem, new_history[1])
-        assert updated_result.content[0].text == "Updated Second"  # type: ignore
-
-        # Other items should be unchanged
-        item1_result = cast(AssistantMessageItem, new_history[0])
-        item3_result = cast(AssistantMessageItem, new_history[2])
-        assert item1_result.content[0].text == "First"  # type: ignore
-        assert item3_result.content[0].text == "Third"  # type: ignore
-
-    def test_insert_new_item_after_previous_item(self):
-        """Test inserting new item after specified previous_item_id"""
-        # Create initial history
-        item1 = AssistantMessageItem(
-            item_id="item_1", role="assistant", content=[AssistantText(text="First")]
-        )
-        item3 = AssistantMessageItem(
-            item_id="item_3", role="assistant", content=[AssistantText(text="Third")]
-        )
-        old_history = [item1, item3]
-
-        # Create new item to insert between them
-        new_item = AssistantMessageItem(
-            item_id="item_2",
-            previous_item_id="item_1",
-            role="assistant",
-            content=[AssistantText(text="Second")],
-        )
-
-        new_history = RealtimeSession._get_new_history(
-            cast(list[RealtimeItem], old_history), new_item
-        )
-
-        # Should have one more item
-        assert len(new_history) == 3
-
-        # Items should be in correct order
-        assert new_history[0].item_id == "item_1"
-        assert new_history[1].item_id == "item_2"
-        assert new_history[2].item_id == "item_3"
-
-        # Content should be correct
-        item2_result = cast(AssistantMessageItem, new_history[1])
-        assert item2_result.content[0].text == "Second"  # type: ignore
-
-    def test_insert_new_item_after_nonexistent_previous_item(self):
-        """Test that item with nonexistent previous_item_id gets added to end"""
-        # Create initial history
-        item1 = AssistantMessageItem(
-            item_id="item_1", role="assistant", content=[AssistantText(text="First")]
-        )
-        old_history = [item1]
-
-        # Create new item with nonexistent previous_item_id
-        new_item = AssistantMessageItem(
-            item_id="item_2",
-            previous_item_id="nonexistent",
-            role="assistant",
-            content=[AssistantText(text="Second")],
-        )
-
-        new_history = RealtimeSession._get_new_history(
-            cast(list[RealtimeItem], old_history), new_item
-        )
-
-        # Should add to end when previous_item_id not found
-        assert len(new_history) == 2
-        assert new_history[0].item_id == "item_1"
-        assert new_history[1].item_id == "item_2"
-
-    def test_add_new_item_to_end_when_no_previous_item_id(self):
-        """Test adding new item to end when no previous_item_id is specified"""
-        # Create initial history
-        item1 = AssistantMessageItem(
-            item_id="item_1", role="assistant", content=[AssistantText(text="First")]
-        )
-        old_history = [item1]
-
-        # Create new item without previous_item_id
-        new_item = AssistantMessageItem(
-            item_id="item_2", role="assistant", content=[AssistantText(text="Second")]
-        )
-
-        new_history = RealtimeSession._get_new_history(
-            cast(list[RealtimeItem], old_history), new_item
-        )
-
-        # Should add to end
-        assert len(new_history) == 2
-        assert new_history[0].item_id == "item_1"
-        assert new_history[1].item_id == "item_2"
-
-    def test_tool_call_item_update_replaces_existing_entry(self):
-        """A completed tool call replaces the in-progress entry it shares an item_id with."""
-        in_progress = RealtimeToolCallItem(
-            item_id="item_1",
-            previous_item_id=None,
-            call_id="call_1",
-            type="function_call",
-            status="in_progress",
-            arguments='{"city": "Oakland"}',
-            name="get_weather",
-            output=None,
-        )
-        completed = RealtimeToolCallItem(
-            item_id="item_1",
-            previous_item_id=None,
-            call_id="call_1",
-            type="function_call",
-            status="completed",
-            arguments='{"city": "Oakland"}',
-            name="get_weather",
-            output="sunny",
-        )
-
-        history = RealtimeSession._get_new_history([], in_progress)
-        history = RealtimeSession._get_new_history(history, completed)
-
-        assert len(history) == 1
-        updated = cast(RealtimeToolCallItem, history[0])
-        assert updated.status == "completed"
-        assert updated.output == "sunny"
-
-    def test_tool_call_item_update_preserves_other_items(self):
-        """Replacing a tool call entry leaves the surrounding history untouched."""
-        before = UserMessageItem(
-            item_id="item_0", role="user", content=[InputText(text="what's the weather?")]
-        )
-        after = AssistantMessageItem(
-            item_id="item_2", role="assistant", content=[AssistantText(text="It is sunny.")]
-        )
-        in_progress = RealtimeToolCallItem(
-            item_id="item_1",
-            previous_item_id=None,
-            call_id="call_1",
-            type="function_call",
-            status="in_progress",
-            arguments="{}",
-            name="get_weather",
-            output=None,
-        )
-        old_history = cast(list[RealtimeItem], [before, in_progress, after])
-
-        completed = in_progress.model_copy(update={"status": "completed", "output": "sunny"})
-        new_history = RealtimeSession._get_new_history(old_history, completed)
-
-        assert [item.item_id for item in new_history] == ["item_0", "item_1", "item_2"]
-        assert new_history[0] == before
-        assert new_history[2] == after
-        assert cast(RealtimeToolCallItem, new_history[1]).output == "sunny"
-
-    def test_add_first_item_to_empty_history(self):
-        """Test adding first item to empty history"""
-        old_history: list[RealtimeItem] = []
-
-        new_item = AssistantMessageItem(
-            item_id="item_1", role="assistant", content=[AssistantText(text="First")]
-        )
-
-        new_history = RealtimeSession._get_new_history(old_history, new_item)
-
-        assert len(new_history) == 1
-        assert new_history[0].item_id == "item_1"
-
-    def test_complex_insertion_scenario(self):
-        """Test complex scenario with multiple insertions and updates"""
-        # Start with items A and C
-        itemA = AssistantMessageItem(
-            item_id="A", role="assistant", content=[AssistantText(text="A")]
-        )
-        itemC = AssistantMessageItem(
-            item_id="C", role="assistant", content=[AssistantText(text="C")]
-        )
-        history: list[RealtimeItem] = [itemA, itemC]
-
-        # Insert B after A
-        itemB = AssistantMessageItem(
-            item_id="B", previous_item_id="A", role="assistant", content=[AssistantText(text="B")]
-        )
-        history = RealtimeSession._get_new_history(history, itemB)
-
-        # Should be A, B, C
-        assert len(history) == 3
-        assert [item.item_id for item in history] == ["A", "B", "C"]
-
-        # Insert D after B
-        itemD = AssistantMessageItem(
-            item_id="D", previous_item_id="B", role="assistant", content=[AssistantText(text="D")]
-        )
-        history = RealtimeSession._get_new_history(history, itemD)
-
-        # Should be A, B, D, C
-        assert len(history) == 4
-        assert [item.item_id for item in history] == ["A", "B", "D", "C"]
-
-        # Update B
-        updated_itemB = AssistantMessageItem(
-            item_id="B", role="assistant", content=[AssistantText(text="Updated B")]
-        )
-        history = RealtimeSession._get_new_history(history, updated_itemB)
-
-        # Should still be A, B, D, C but B is updated
-        assert len(history) == 4
-        assert [item.item_id for item in history] == ["A", "B", "D", "C"]
-        itemB_result = cast(AssistantMessageItem, history[1])
-        assert itemB_result.content[0].text == "Updated B"  # type: ignore
 
 
 # Test 3: Tool call execution flow (_handle_tool_call method)
@@ -2469,43 +1764,6 @@ class TestToolCallExecution:
             await session.__aexit__(None, None, None)
 
     @pytest.mark.asyncio
-    async def test_approval_resume_uses_pending_initial_settings_dispatch_snapshot(
-        self, mock_model
-    ):
-        approved_tool = _named_function_tool(
-            "approval_tool",
-            "approved implementation",
-            needs_approval=True,
-        )
-        replacement_tool = _named_function_tool("approval_tool", "replacement implementation")
-        initial_agent = RealtimeAgent(name="initial", tools=[], handoffs=[])
-        replacement_agent = RealtimeAgent(name="replacement", tools=[replacement_tool], handoffs=[])
-        session = RealtimeSession(
-            mock_model,
-            initial_agent,
-            None,
-            model_config={"initial_model_settings": {"tools": [approved_tool]}},
-            run_config={"async_tool_calls": False},
-        )
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="approval_tool",
-            call_id="call_pending_snapshot",
-            arguments="{}",
-        )
-
-        await session.__aenter__()
-        try:
-            await session._handle_tool_call(tool_call_event)
-            assert list(session._pending_tool_calls) == [tool_call_event.call_id]
-
-            await session.update_agent(replacement_agent)
-            await session.approve_tool_call(tool_call_event.call_id)
-
-            assert _sent_tool_output_strings(mock_model) == ["approved implementation"]
-        finally:
-            await session.__aexit__(None, None, None)
-
-    @pytest.mark.asyncio
     async def test_async_tool_call_uses_event_initial_settings_dispatch_snapshot(
         self, mock_model, monkeypatch
     ):
@@ -2569,195 +1827,6 @@ class TestToolCallExecution:
         await session._handle_tool_call(tool_call_event)
         await session._handle_tool_call(tool_call_event)
 
-        mock_function_tool.on_invoke_tool.assert_called_once()
-        assert len(mock_model.sent_tool_outputs) == 1
-
-    @pytest.mark.asyncio
-    async def test_approved_function_tool_failure_replay_does_not_rerun(
-        self, mock_model, mock_agent, mock_function_tool
-    ):
-        mock_function_tool.needs_approval = True
-        mock_function_tool.on_invoke_tool.side_effect = RuntimeError("failed after side effect")
-        mock_agent.get_all_tools.return_value = [mock_function_tool]
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={"async_tool_calls": False},
-        )
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_failed", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-        with pytest.raises(RuntimeError, match="failed after side effect"):
-            await session.approve_tool_call(tool_call_event.call_id)
-
-        with pytest.raises(ModelBehaviorError, match="already executed"):
-            await session._handle_tool_call(tool_call_event)
-
-        mock_function_tool.on_invoke_tool.assert_awaited_once()
-        assert len(mock_model.sent_tool_outputs) == 0
-
-    @pytest.mark.parametrize("always", [False, True], ids=["per-call", "sticky"])
-    @pytest.mark.parametrize("changed_field", ["arguments", "tool_name"])
-    @pytest.mark.asyncio
-    async def test_function_tool_send_failure_retries_cached_output_without_rerun(
-        self,
-        mock_agent,
-        mock_function_tool,
-        always: bool,
-        changed_field: str,
-    ):
-        """An approved call should retry cached output only for the same invocation."""
-
-        class FailingToolOutputModel(RecordingRealtimeModel):
-            def __init__(self):
-                super().__init__()
-                self.fail_next_tool_output = True
-
-            async def send_event(self, event):
-                if isinstance(event, RealtimeModelSendToolOutput) and self.fail_next_tool_output:
-                    self.fail_next_tool_output = False
-                    raise RuntimeError("send failed")
-                await super().send_event(event)
-
-        mock_function_tool.needs_approval = True
-        mock_agent.get_all_tools.return_value = [mock_function_tool]
-        mock_model = FailingToolOutputModel()
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={"async_tool_calls": False},
-        )
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_retry_output", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-        with pytest.raises(RuntimeError, match="send failed"):
-            await session.approve_tool_call(tool_call_event.call_id, always=always)
-
-        mock_function_tool.on_invoke_tool.assert_called_once()
-        assert len(mock_model.sent_tool_outputs) == 0
-
-        changed_event = RealtimeModelToolCallEvent(
-            name="other_function" if changed_field == "tool_name" else tool_call_event.name,
-            call_id=tool_call_event.call_id,
-            arguments=(
-                tool_call_event.arguments if changed_field == "tool_name" else '{"changed":true}'
-            ),
-        )
-        with pytest.raises(ModelBehaviorError, match="unique call ID"):
-            await session._handle_tool_call(changed_event)
-        await session._handle_tool_call(tool_call_event)
-
-        mock_function_tool.on_invoke_tool.assert_called_once()
-        assert len(mock_model.sent_tool_outputs) == 1
-
-    @pytest.mark.asyncio
-    async def test_tool_end_cancellation_after_output_send_does_not_resend(
-        self, mock_model, mock_agent, mock_function_tool
-    ) -> None:
-        """Provider delivery commits the output before local end-event publication."""
-        mock_agent.get_all_tools.return_value = [mock_function_tool]
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={"async_tool_calls": False},
-        )
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function",
-            call_id="call_tool_end_cancelled",
-            arguments="{}",
-        )
-        original_put_event_nowait = session._put_event_nowait
-
-        def cancel_tool_end(event: Any) -> bool:
-            if isinstance(event, RealtimeToolEnd):
-                raise asyncio.CancelledError
-            return original_put_event_nowait(event)
-
-        session._put_event_nowait = cancel_tool_end  # type: ignore[method-assign]
-        with pytest.raises(asyncio.CancelledError):
-            await session._handle_tool_call(tool_call_event)
-
-        invocation = session._context_wrapper._tool_invocations[tool_call_event.call_id]
-        assert invocation.executed is True
-        assert invocation.completed is True
-        assert tool_call_event.call_id not in session._pending_tool_outputs
-        mock_function_tool.on_invoke_tool.assert_called_once()
-        assert len(mock_model.sent_tool_outputs) == 1
-
-        session._put_event_nowait = original_put_event_nowait  # type: ignore[method-assign]
-        await session._handle_tool_call(tool_call_event)
-
-        mock_function_tool.on_invoke_tool.assert_called_once()
-        assert len(mock_model.sent_tool_outputs) == 1
-
-    @pytest.mark.parametrize("always", [False, True], ids=["per-call", "sticky"])
-    @pytest.mark.parametrize("changed_field", ["arguments", "tool_name"])
-    @pytest.mark.asyncio
-    async def test_async_function_tool_send_failure_retries_cached_output_without_rerun(
-        self,
-        mock_agent,
-        mock_function_tool,
-        always: bool,
-        changed_field: str,
-    ):
-        """The async approval path should bind retries to the original invocation."""
-
-        class FailingToolOutputModel(RecordingRealtimeModel):
-            def __init__(self):
-                super().__init__()
-                self.fail_next_tool_output = True
-
-            async def send_event(self, event):
-                if isinstance(event, RealtimeModelSendToolOutput) and self.fail_next_tool_output:
-                    self.fail_next_tool_output = False
-                    raise RuntimeError("send failed")
-                await super().send_event(event)
-
-        mock_function_tool.needs_approval = True
-        mock_agent.get_all_tools.return_value = [mock_function_tool]
-        mock_model = FailingToolOutputModel()
-        session = RealtimeSession(mock_model, mock_agent, None)
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_async_retry_output", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-        await session.approve_tool_call(tool_call_event.call_id, always=always)
-        tool_call_tasks = list(session._tool_call_tasks)
-        assert len(tool_call_tasks) == 1
-        task_results = await asyncio.gather(*tool_call_tasks, return_exceptions=True)
-        await asyncio.sleep(0)
-
-        assert len(task_results) == 1
-        assert isinstance(task_results[0], RuntimeError)
-        assert session._stored_exception is None
-        assert tool_call_event.call_id in session._pending_tool_outputs
-        mock_function_tool.on_invoke_tool.assert_called_once()
-        assert len(mock_model.sent_tool_outputs) == 0
-
-        changed_event = RealtimeModelToolCallEvent(
-            name="other_function" if changed_field == "tool_name" else tool_call_event.name,
-            call_id=tool_call_event.call_id,
-            arguments=(
-                tool_call_event.arguments if changed_field == "tool_name" else '{"changed":true}'
-            ),
-        )
-        with pytest.raises(ModelBehaviorError, match="unique call ID"):
-            await session._handle_tool_call(changed_event)
-        await session.on_event(tool_call_event)
-        tool_call_tasks = list(session._tool_call_tasks)
-        assert len(tool_call_tasks) == 1
-        await asyncio.gather(*tool_call_tasks)
-
-        assert session._stored_exception is None
-        assert tool_call_event.call_id not in session._pending_tool_outputs
         mock_function_tool.on_invoke_tool.assert_called_once()
         assert len(mock_model.sent_tool_outputs) == 1
 
@@ -3099,154 +2168,6 @@ class TestToolCallExecution:
         mock_function_tool.on_invoke_tool.assert_not_called()
 
     @pytest.mark.asyncio
-    async def test_function_tool_needs_approval_emits_event(
-        self, mock_model, mock_agent, mock_function_tool
-    ):
-        """Tools marked as needs_approval should pause and emit an approval request."""
-        mock_function_tool.needs_approval = True
-        mock_agent.get_all_tools.return_value = [mock_function_tool]
-
-        session = RealtimeSession(mock_model, mock_agent, None)
-
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_needs_approval", arguments='{"param": "value"}'
-        )
-
-        await session._handle_tool_call(tool_call_event)
-
-        assert tool_call_event.call_id in session._pending_tool_calls
-        assert mock_function_tool.on_invoke_tool.call_count == 0
-
-        approval_event = await session._event_queue.get()
-        assert isinstance(approval_event, RealtimeToolApprovalRequired)
-        assert approval_event.call_id == tool_call_event.call_id
-        assert approval_event.tool == mock_function_tool
-
-    @pytest.mark.parametrize(
-        "arguments",
-        [
-            "",
-            '{"subject": "refund"',
-            "null",
-            "[]",
-            '{"amount": NaN}',
-            '{"amount": Infinity}',
-            '{"amount": -Infinity}',
-        ],
-    )
-    @pytest.mark.asyncio
-    async def test_callable_function_approval_fails_closed_for_invalid_arguments(
-        self, mock_model, arguments: str
-    ) -> None:
-        approval_inputs: list[dict[str, Any]] = []
-        tool_inputs: list[str] = []
-
-        async def needs_approval(_ctx: Any, params: dict[str, Any], _call_id: str) -> bool:
-            approval_inputs.append(params)
-            return False
-
-        async def invoke_tool(_ctx: ToolContext[Any], raw_arguments: str) -> str:
-            tool_inputs.append(raw_arguments)
-            return "sent"
-
-        tool = FunctionTool(
-            name="send_email",
-            description="Send an email.",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=invoke_tool,
-            needs_approval=needs_approval,
-        )
-        agent = RealtimeAgent(name="agent", tools=[tool])
-        session = RealtimeSession(mock_model, agent, None, run_config={"async_tool_calls": False})
-        tool_call_event = RealtimeModelToolCallEvent(
-            name=tool.name,
-            call_id="call-invalid",
-            arguments=arguments,
-        )
-
-        await session._handle_tool_call(tool_call_event)
-
-        assert tool_call_event.call_id in session._pending_tool_calls
-        assert approval_inputs == []
-        assert tool_inputs == []
-        approval_event = await session._event_queue.get()
-        assert isinstance(approval_event, RealtimeToolApprovalRequired)
-
-    @pytest.mark.asyncio
-    async def test_callable_function_approval_receives_valid_object_arguments(
-        self, mock_model
-    ) -> None:
-        approval_inputs: list[dict[str, Any]] = []
-        tool_inputs: list[str] = []
-
-        async def needs_approval(_ctx: Any, params: dict[str, Any], _call_id: str) -> bool:
-            approval_inputs.append(params)
-            return False
-
-        async def invoke_tool(_ctx: ToolContext[Any], raw_arguments: str) -> str:
-            tool_inputs.append(raw_arguments)
-            return "sent"
-
-        tool = FunctionTool(
-            name="send_email",
-            description="Send an email.",
-            params_json_schema={"type": "object", "properties": {"subject": {"type": "string"}}},
-            on_invoke_tool=invoke_tool,
-            needs_approval=needs_approval,
-        )
-        agent = RealtimeAgent(name="agent", tools=[tool])
-        session = RealtimeSession(mock_model, agent, None, run_config={"async_tool_calls": False})
-        arguments = '{"subject": "status update"}'
-        tool_call_event = RealtimeModelToolCallEvent(
-            name=tool.name,
-            call_id="call-valid",
-            arguments=arguments,
-        )
-
-        await session._handle_tool_call(tool_call_event)
-
-        assert approval_inputs == [{"subject": "status update"}]
-        assert tool_inputs == [arguments]
-        assert tool_call_event.call_id not in session._pending_tool_calls
-
-    @pytest.mark.asyncio
-    async def test_tool_input_guardrail_rejects_before_realtime_function_execution(
-        self, mock_model
-    ):
-        """Tool input guardrails should run before regular realtime function tool execution."""
-        executed = False
-
-        @tool_input_guardrail
-        def reject_guardrail(_data: ToolInputGuardrailData) -> ToolGuardrailFunctionOutput:
-            return ToolGuardrailFunctionOutput.reject_content("blocked before execution")
-
-        async def invoke_tool(_ctx: ToolContext[Any], _arguments: str) -> str:
-            nonlocal executed
-            executed = True
-            return "ok"
-
-        guarded_tool = FunctionTool(
-            name="test_function",
-            description="guarded",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=invoke_tool,
-            tool_input_guardrails=[reject_guardrail],
-        )
-        agent = RealtimeAgent(name="agent", tools=[guarded_tool])
-        session = RealtimeSession(mock_model, agent, None, run_config={"async_tool_calls": False})
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_guardrail_reject", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-
-        assert executed is False
-        assert len(mock_model.sent_tool_outputs) == 1
-        _sent_call, sent_output, start_response = mock_model.sent_tool_outputs[0]
-        assert sent_output == "blocked before execution"
-        assert start_response is True
-
-    @pytest.mark.asyncio
     async def test_realtime_tool_contexts_share_session_tool_state(self, mock_model):
         """Realtime guardrails and callbacks receive the session-owned tool state."""
         observed_contexts: list[ToolContext[Any]] = []
@@ -3282,183 +2203,6 @@ class TestToolCallExecution:
         for context in observed_contexts:
             assert context._approvals is session._context_wrapper._approvals
             assert context._tool_invocations is session._context_wrapper._tool_invocations
-
-    @pytest.mark.asyncio
-    async def test_realtime_pending_approval_skips_tool_input_guardrails_by_default(
-        self, mock_model
-    ):
-        guardrail_runs = 0
-
-        @tool_input_guardrail
-        def count_guardrail(_data: ToolInputGuardrailData) -> ToolGuardrailFunctionOutput:
-            nonlocal guardrail_runs
-            guardrail_runs += 1
-            return ToolGuardrailFunctionOutput.allow()
-
-        async def invoke_tool(_ctx: ToolContext[Any], _arguments: str) -> str:
-            return "ok"
-
-        guarded_tool = FunctionTool(
-            name="test_function",
-            description="guarded",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=invoke_tool,
-            needs_approval=True,
-            tool_input_guardrails=[count_guardrail],
-        )
-        agent = RealtimeAgent(name="agent", tools=[guarded_tool])
-        session = RealtimeSession(mock_model, agent, None, run_config={"async_tool_calls": False})
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_guardrail_pending", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-
-        assert tool_call_event.call_id in session._pending_tool_calls
-        assert guardrail_runs == 0
-
-    @pytest.mark.asyncio
-    async def test_realtime_pre_approval_tool_input_guardrail_rejects_pending_approval(
-        self, mock_model
-    ):
-        executed = False
-
-        @tool_input_guardrail
-        def reject_guardrail(_data: ToolInputGuardrailData) -> ToolGuardrailFunctionOutput:
-            return ToolGuardrailFunctionOutput.reject_content("blocked before approval")
-
-        async def invoke_tool(_ctx: ToolContext[Any], _arguments: str) -> str:
-            nonlocal executed
-            executed = True
-            return "ok"
-
-        guarded_tool = FunctionTool(
-            name="test_function",
-            description="guarded",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=invoke_tool,
-            needs_approval=True,
-            tool_input_guardrails=[reject_guardrail],
-        )
-        agent = RealtimeAgent(name="agent", tools=[guarded_tool])
-        session = RealtimeSession(
-            mock_model,
-            agent,
-            None,
-            run_config={
-                "async_tool_calls": False,
-                "tool_execution": {"pre_approval_tool_input_guardrails": True},
-            },
-        )
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_pre_approval_reject", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-
-        assert executed is False
-        assert tool_call_event.call_id not in session._pending_tool_calls
-        assert len(mock_model.sent_tool_outputs) == 1
-        _sent_call, sent_output, start_response = mock_model.sent_tool_outputs[0]
-        assert sent_output == "blocked before approval"
-        assert start_response is True
-
-    @pytest.mark.asyncio
-    async def test_realtime_pre_approval_tool_input_guardrails_rerun_after_approval(
-        self, mock_model
-    ):
-        guardrail_runs = 0
-        executed = 0
-
-        @tool_input_guardrail
-        def count_guardrail(_data: ToolInputGuardrailData) -> ToolGuardrailFunctionOutput:
-            nonlocal guardrail_runs
-            guardrail_runs += 1
-            return ToolGuardrailFunctionOutput.allow()
-
-        async def invoke_tool(_ctx: ToolContext[Any], _arguments: str) -> str:
-            nonlocal executed
-            executed += 1
-            return "ok"
-
-        guarded_tool = FunctionTool(
-            name="test_function",
-            description="guarded",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=invoke_tool,
-            needs_approval=True,
-            tool_input_guardrails=[count_guardrail],
-        )
-        agent = RealtimeAgent(name="agent", tools=[guarded_tool])
-        session = RealtimeSession(
-            mock_model,
-            agent,
-            None,
-            run_config={
-                "async_tool_calls": False,
-                "tool_execution": {"pre_approval_tool_input_guardrails": True},
-            },
-        )
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_pre_approval_rerun", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-        assert guardrail_runs == 1
-        assert executed == 0
-
-        await session.approve_tool_call(tool_call_event.call_id)
-
-        assert guardrail_runs == 2
-        assert executed == 1
-        assert len(mock_model.sent_tool_outputs) == 1
-        _sent_call, sent_output, start_response = mock_model.sent_tool_outputs[0]
-        assert sent_output == "ok"
-        assert start_response is True
-
-    @pytest.mark.asyncio
-    async def test_duplicate_pending_approval_call_id_is_ignored_and_approval_runs_once(
-        self, mock_model, mock_agent, mock_function_tool
-    ):
-        """A duplicate approval-gated call should not enqueue another approval or run twice."""
-        mock_function_tool.needs_approval = True
-        mock_agent.get_all_tools.return_value = [mock_function_tool]
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={"async_tool_calls": False},
-        )
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_duplicate_approval", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-        await session._handle_tool_call(tool_call_event)
-
-        changed_event = RealtimeModelToolCallEvent(
-            name="test_function",
-            call_id=tool_call_event.call_id,
-            arguments='{"changed":true}',
-        )
-        with pytest.raises(ModelBehaviorError, match="unique call ID"):
-            await session._handle_tool_call(changed_event)
-
-        assert list(session._pending_tool_calls) == [tool_call_event.call_id]
-        approval_events = []
-        while not session._event_queue.empty():
-            event = await session._event_queue.get()
-            if isinstance(event, RealtimeToolApprovalRequired):
-                approval_events.append(event)
-        assert len(approval_events) == 1
-
-        await session.approve_tool_call(tool_call_event.call_id)
-        await session._handle_tool_call(tool_call_event)
-        with pytest.raises(ModelBehaviorError, match="unique call ID"):
-            await session._handle_tool_call(changed_event)
-
-        mock_function_tool.on_invoke_tool.assert_called_once()
-        assert len(mock_model.sent_tool_outputs) == 1
 
     @pytest.mark.asyncio
     async def test_changed_realtime_call_id_fails_while_dispatch_resolution_is_pending(
@@ -3634,461 +2378,6 @@ class TestToolCallExecution:
         assert executed == ["{}"]
 
     @pytest.mark.asyncio
-    async def test_approve_pending_tool_call_runs_tool(
-        self, mock_model, mock_agent, mock_function_tool
-    ):
-        """Approving a pending tool call should resume execution."""
-        mock_function_tool.needs_approval = True
-        mock_agent.get_all_tools.return_value = [mock_function_tool]
-
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={"async_tool_calls": False},
-        )
-
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_approve", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-        await session.approve_tool_call(tool_call_event.call_id)
-
-        assert mock_function_tool.on_invoke_tool.call_count == 1
-        assert len(mock_model.sent_tool_outputs) == 1
-        assert session._pending_tool_calls == {}
-
-        events = []
-        while not session._event_queue.empty():
-            events.append(await session._event_queue.get())
-
-        assert any(isinstance(ev, RealtimeToolStart) for ev in events)
-        assert any(isinstance(ev, RealtimeToolEnd) for ev in events)
-
-    @pytest.mark.asyncio
-    async def test_async_approve_pending_tool_call_reserves_call_id_before_task_runs(
-        self, mock_model
-    ):
-        """A duplicate event after approval should not outrun the approved async task."""
-        approved_calls: list[str] = []
-        duplicate_calls: list[str] = []
-
-        async def invoke_approved_tool(_ctx: ToolContext[Any], _arguments: str) -> str:
-            approved_calls.append("approved")
-            return "approved_result"
-
-        async def invoke_duplicate_tool(_ctx: ToolContext[Any], _arguments: str) -> str:
-            duplicate_calls.append("duplicate")
-            return "duplicate_result"
-
-        approved_tool = FunctionTool(
-            name="test_function",
-            description="approved",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=invoke_approved_tool,
-            needs_approval=True,
-        )
-        duplicate_tool = FunctionTool(
-            name="test_function",
-            description="duplicate",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=invoke_duplicate_tool,
-            needs_approval=False,
-        )
-        approved_agent = RealtimeAgent(name="approved_agent", tools=[approved_tool])
-        duplicate_agent = RealtimeAgent(name="duplicate_agent", tools=[duplicate_tool])
-        session = RealtimeSession(mock_model, approved_agent, None)
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_async_approval_race", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-        await session.approve_tool_call(tool_call_event.call_id)
-
-        assert tool_call_event.call_id in session._active_tool_invocations
-        await session._handle_tool_call(tool_call_event, agent_snapshot=duplicate_agent)
-
-        tool_call_tasks = list(session._tool_call_tasks)
-        assert len(tool_call_tasks) == 1
-        await asyncio.gather(*tool_call_tasks)
-
-        assert approved_calls == ["approved"]
-        assert duplicate_calls == []
-        assert len(mock_model.sent_tool_outputs) == 1
-        _sent_call, sent_output, _start_response = mock_model.sent_tool_outputs[0]
-        assert sent_output == "approved_result"
-
-    @pytest.mark.asyncio
-    async def test_always_approve_namespaced_tool_call_does_not_approve_bare_tool(self, mock_model):
-        """Always approval should stay scoped to the namespaced tool key."""
-        tool_calls: list[str] = []
-
-        async def invoke_tool(_ctx: ToolContext[Any], _arguments: str) -> str:
-            tool_calls.append("called")
-            return "account"
-
-        namespaced_tool = tool_namespace(
-            name="crm",
-            description="CRM tools",
-            tools=[
-                FunctionTool(
-                    name="lookup_account",
-                    description="Look up account",
-                    params_json_schema={"type": "object", "properties": {}},
-                    on_invoke_tool=invoke_tool,
-                    needs_approval=True,
-                )
-            ],
-        )[0]
-        bare_tool = FunctionTool(
-            name="lookup_account",
-            description="Look up account",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=invoke_tool,
-            needs_approval=True,
-        )
-        namespaced_agent = RealtimeAgent(name="crm_agent", tools=[namespaced_tool])
-        bare_agent = RealtimeAgent(name="bare_agent", tools=[bare_tool])
-
-        session = RealtimeSession(
-            mock_model,
-            namespaced_agent,
-            None,
-            run_config={"async_tool_calls": False},
-        )
-
-        first_call = RealtimeModelToolCallEvent(
-            name="lookup_account", call_id="call_first", arguments="{}"
-        )
-        second_call = RealtimeModelToolCallEvent(
-            name="lookup_account", call_id="call_second", arguments="{}"
-        )
-
-        await session._handle_tool_call(first_call)
-        await session.approve_tool_call(first_call.call_id, always=True)
-        await session._handle_tool_call(second_call, agent_snapshot=bare_agent)
-
-        assert (
-            session._context_wrapper.get_approval_status(
-                "lookup_account",
-                second_call.call_id,
-            )
-            is None
-        )
-        assert "crm.lookup_account" in session._context_wrapper._approvals
-        assert "lookup_account" not in session._context_wrapper._approvals
-        assert sorted(session._pending_tool_calls) == [second_call.call_id]
-        assert len(mock_model.sent_tool_outputs) == 1
-        assert tool_calls == ["called"]
-
-    @pytest.mark.asyncio
-    async def test_reject_pending_tool_call_sends_rejection_output(
-        self, mock_model, mock_agent, mock_function_tool
-    ):
-        """Rejecting a pending tool call should notify the model and skip execution."""
-        mock_function_tool.needs_approval = True
-        mock_agent.get_all_tools.return_value = [mock_function_tool]
-
-        session = RealtimeSession(mock_model, mock_agent, None)
-
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_reject", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-        await session.reject_tool_call(tool_call_event.call_id)
-        await session._handle_tool_call(tool_call_event)
-
-        assert mock_function_tool.on_invoke_tool.call_count == 0
-        assert len(mock_model.sent_tool_outputs) == 1
-        _sent_call, sent_output, start_response = mock_model.sent_tool_outputs[0]
-        assert sent_output == REJECTION_MESSAGE
-        assert start_response is True
-        assert session._pending_tool_calls == {}
-
-        events = []
-        while not session._event_queue.empty():
-            events.append(await session._event_queue.get())
-
-        assert any(
-            isinstance(ev, RealtimeToolEnd) and ev.output == REJECTION_MESSAGE for ev in events
-        )
-
-    @pytest.mark.asyncio
-    async def test_reject_pending_tool_call_reserves_call_id_before_sending(
-        self, mock_agent, mock_function_tool
-    ):
-        """A duplicate event during rejection output sending should not emit a second output."""
-
-        class BlockingToolOutputModel(RecordingRealtimeModel):
-            def __init__(self):
-                super().__init__()
-                self.started = asyncio.Event()
-                self.release = asyncio.Event()
-                self.block_next_tool_output = True
-
-            async def send_event(self, event):
-                if isinstance(event, RealtimeModelSendToolOutput) and self.block_next_tool_output:
-                    self.block_next_tool_output = False
-                    self.started.set()
-                    await self.release.wait()
-                await super().send_event(event)
-
-        mock_function_tool.needs_approval = True
-        mock_agent.get_all_tools.return_value = [mock_function_tool]
-        mock_model = BlockingToolOutputModel()
-        session = RealtimeSession(mock_model, mock_agent, None)
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_reject_race", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-        reject_task = asyncio.create_task(session.reject_tool_call(tool_call_event.call_id))
-        await asyncio.wait_for(mock_model.started.wait(), timeout=1)
-
-        await session._handle_tool_call(tool_call_event)
-
-        mock_model.release.set()
-        await reject_task
-
-        assert len(mock_model.sent_tool_outputs) == 1
-
-    @pytest.mark.asyncio
-    async def test_reject_pending_tool_call_uses_run_level_formatter(
-        self, mock_model, mock_agent, mock_function_tool
-    ):
-        """Rejecting a pending tool call should use the run-level formatter output."""
-        mock_function_tool.needs_approval = True
-        mock_agent.get_all_tools.return_value = [mock_function_tool]
-
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={
-                "tool_error_formatter": (
-                    lambda args: f"run-level {args.tool_name} denied ({args.call_id})"
-                )
-            },
-        )
-
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_reject_custom", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-        await session.reject_tool_call(tool_call_event.call_id)
-
-        _sent_call, sent_output, start_response = mock_model.sent_tool_outputs[0]
-        assert sent_output == "run-level test_function denied (call_reject_custom)"
-        assert start_response is True
-
-        events = []
-        while not session._event_queue.empty():
-            events.append(await session._event_queue.get())
-
-        assert any(
-            isinstance(ev, RealtimeToolEnd)
-            and ev.output == "run-level test_function denied (call_reject_custom)"
-            for ev in events
-        )
-
-    @pytest.mark.asyncio
-    async def test_rejection_formatter_error_is_redacted(
-        self, monkeypatch, mock_model, mock_agent, mock_function_tool
-    ):
-        monkeypatch.setattr(_debug, "DONT_LOG_TOOL_DATA", True)
-
-        def fail_formatter(_args):
-            raise ValueError("SECRET_REALTIME_TOOL_FORMATTER")
-
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={"tool_error_formatter": fail_formatter},
-        )
-
-        with patch("agents.realtime.session.logger") as mock_logger:
-            message = await session._resolve_approval_rejection_message(
-                tool=mock_function_tool,
-                call_id="call_reject_error",
-            )
-
-        assert message
-        mock_logger.error.assert_called_once_with("%s", "Tool error formatter failed", stacklevel=3)
-
-    @pytest.mark.asyncio
-    async def test_cancelled_rejection_formatter_leaves_invocation_executed(
-        self, mock_model, mock_agent
-    ):
-        formatter_entered = asyncio.Event()
-
-        @function_tool
-        def approval_tool() -> str:
-            return "done"
-
-        async def blocking_formatter(_args):
-            formatter_entered.set()
-            await asyncio.Event().wait()
-            return "rejected"
-
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={"tool_error_formatter": blocking_formatter},
-        )
-        tool_call = RealtimeModelToolCallEvent(
-            name=approval_tool.name,
-            call_id="call_rejected_cancelled",
-            arguments="{}",
-        )
-        canonical_call = session._build_tool_approval_item(  # noqa: SLF001
-            approval_tool,
-            tool_call,
-            mock_agent,
-        ).raw_item
-        lookup_key = get_function_tool_lookup_key_for_tool(approval_tool)
-        assert session._context_wrapper._tool_invocation_status(  # noqa: SLF001
-            canonical_call,
-            tool_lookup_key=lookup_key,
-        ) == (("function_call", "call_rejected_cancelled"), False, False)
-
-        task = asyncio.create_task(
-            session._resolve_approval_rejection_message(  # noqa: SLF001
-                tool=approval_tool,
-                call_id=tool_call.call_id,
-                tool_call=canonical_call,
-            )
-        )
-        await formatter_entered.wait()
-        task.cancel()
-        with pytest.raises(asyncio.CancelledError):
-            await task
-
-        assert session._context_wrapper._tool_invocation_status(  # noqa: SLF001
-            canonical_call,
-            tool_lookup_key=lookup_key,
-        ) == (("function_call", "call_rejected_cancelled"), False, True)
-
-    @pytest.mark.asyncio
-    async def test_reject_pending_tool_call_prefers_explicit_message(
-        self, mock_model, mock_agent, mock_function_tool
-    ):
-        """Rejecting a pending tool call should prefer the explicit rejection message."""
-        mock_function_tool.needs_approval = True
-        mock_agent.get_all_tools.return_value = [mock_function_tool]
-
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={
-                "tool_error_formatter": (
-                    lambda args: f"run-level {args.tool_name} denied ({args.call_id})"
-                )
-            },
-        )
-
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_reject_explicit", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-        await session.reject_tool_call(
-            tool_call_event.call_id,
-            rejection_message="explicit rejection message",
-        )
-
-        _sent_call, sent_output, start_response = mock_model.sent_tool_outputs[0]
-        assert sent_output == "explicit rejection message"
-        assert start_response is True
-
-        events = []
-        while not session._event_queue.empty():
-            events.append(await session._event_queue.get())
-
-        assert any(
-            isinstance(ev, RealtimeToolEnd) and ev.output == "explicit rejection message"
-            for ev in events
-        )
-
-    @pytest.mark.asyncio
-    async def test_always_reject_namespaced_tool_call_reuses_explicit_message(self, mock_model):
-        """Always rejection should reuse explicit messages through the qualified tool key."""
-        tool_calls: list[str] = []
-
-        async def invoke_tool(_ctx: ToolContext[Any], _arguments: str) -> str:
-            tool_calls.append("called")
-            return "account"
-
-        namespaced_tool = tool_namespace(
-            name="crm",
-            description="CRM tools",
-            tools=[
-                FunctionTool(
-                    name="lookup_account",
-                    description="Look up account",
-                    params_json_schema={"type": "object", "properties": {}},
-                    on_invoke_tool=invoke_tool,
-                    needs_approval=True,
-                )
-            ],
-        )[0]
-        agent = RealtimeAgent(name="crm_agent", tools=[namespaced_tool])
-        session = RealtimeSession(mock_model, agent, None)
-
-        first_call = RealtimeModelToolCallEvent(
-            name="lookup_account", call_id="call_reject_first", arguments="{}"
-        )
-        second_call = RealtimeModelToolCallEvent(
-            name="lookup_account", call_id="call_reject_second", arguments="{}"
-        )
-
-        await session._handle_tool_call(first_call)
-        await session.reject_tool_call(
-            first_call.call_id,
-            always=True,
-            rejection_message="explicit crm rejection",
-        )
-        await session._handle_tool_call(second_call)
-
-        assert "crm.lookup_account" in session._context_wrapper._approvals
-        assert "lookup_account" not in session._context_wrapper._approvals
-        assert session._pending_tool_calls == {}
-        assert [output for _call, output, _start in mock_model.sent_tool_outputs] == [
-            "explicit crm rejection",
-            "explicit crm rejection",
-        ]
-        assert tool_calls == []
-
-    @pytest.mark.asyncio
-    async def test_sticky_rejection_does_not_bind_duplicate_call_id_payload(
-        self, mock_model, mock_agent, mock_function_tool
-    ):
-        mock_function_tool.needs_approval = True
-        mock_agent.get_all_tools.return_value = [mock_function_tool]
-        session = RealtimeSession(mock_model, mock_agent, None)
-        first_call = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call-sticky-reject", arguments="{}"
-        )
-        changed_call = RealtimeModelToolCallEvent(
-            name="test_function",
-            call_id=first_call.call_id,
-            arguments='{"changed":true}',
-        )
-
-        await session._handle_tool_call(first_call)
-        await session.reject_tool_call(first_call.call_id, always=True)
-        with pytest.raises(ModelBehaviorError, match="unique call ID"):
-            await session._handle_tool_call(changed_call)
-
-        mock_function_tool.on_invoke_tool.assert_not_called()
-        assert len(mock_model.sent_tool_outputs) == 1
-
-    @pytest.mark.asyncio
     async def test_changed_completed_non_approval_call_id_fails_before_execution(
         self, mock_model, mock_agent, mock_function_tool
     ):
@@ -4183,41 +2472,6 @@ class TestToolCallExecution:
         assert any(isinstance(result, ModelBehaviorError) for result in results)
         assert function_calls == ["function"]
 
-    @pytest.mark.asyncio
-    async def test_pending_function_output_rejects_handoff_role_reuse(self):
-        class FailingToolOutputModel(RecordingRealtimeModel):
-            async def send_event(self, event):
-                if isinstance(event, RealtimeModelSendToolOutput):
-                    raise RuntimeError("send failed")
-                await super().send_event(event)
-
-        function_callback = AsyncMock(return_value="function result")
-        function_tool = FunctionTool(
-            name="route",
-            description="Run a function.",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=function_callback,
-        )
-        function_agent = RealtimeAgent(name="function", tools=[function_tool])
-        target = RealtimeAgent(name="target")
-        route_name = Handoff.default_tool_name(target)
-        function_tool.name = route_name
-        handoff_agent = RealtimeAgent(name="handoff", handoffs=[target])
-        session = RealtimeSession(
-            FailingToolOutputModel(),
-            function_agent,
-            None,
-            run_config={"async_tool_calls": False},
-        )
-        event = RealtimeModelToolCallEvent(name=route_name, call_id="shared", arguments="{}")
-
-        with pytest.raises(RuntimeError, match="send failed"):
-            await session._handle_tool_call(event)
-        with pytest.raises(ModelBehaviorError, match="unique call ID"):
-            await session._handle_tool_call(event, agent_snapshot=handoff_agent)
-
-        function_callback.assert_awaited_once()
-
     @pytest.mark.parametrize(
         "failure",
         [RuntimeError("settings failed"), asyncio.CancelledError()],
@@ -4263,49 +2517,6 @@ class TestToolCallExecution:
         callback.assert_awaited_once()
 
     @pytest.mark.asyncio
-    async def test_async_exact_function_retry_after_serialization_failure_does_not_repeat_callback(
-        self,
-        mock_model,
-    ):
-        callback = AsyncMock(return_value={"result": "ok"})
-        tool = FunctionTool(
-            name="run_function",
-            description="Run a function.",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=callback,
-        )
-        agent = RealtimeAgent(name="agent", tools=[tool])
-        session = RealtimeSession(mock_model, agent, None)
-        event = RealtimeModelToolCallEvent(
-            name=tool.name,
-            call_id="shared",
-            arguments="{}",
-        )
-
-        with patch(
-            "agents.realtime.session._serialize_tool_output",
-            side_effect=RuntimeError("serialization failed"),
-        ):
-            await session.on_event(event)
-            first_results = await asyncio.gather(
-                *list(session._tool_call_tasks),
-                return_exceptions=True,
-            )
-
-        await session.on_event(event)
-        retry_results = await asyncio.gather(
-            *list(session._tool_call_tasks),
-            return_exceptions=True,
-        )
-
-        assert any(
-            isinstance(result, RuntimeError) and str(result) == "serialization failed"
-            for result in first_results
-        )
-        assert any(isinstance(result, ModelBehaviorError) for result in retry_results)
-        callback.assert_awaited_once()
-
-    @pytest.mark.asyncio
     async def test_empty_handoff_call_id_fails_before_callback(self, mock_model):
         target = RealtimeAgent(name="target")
         callback = AsyncMock(return_value=target)
@@ -4327,186 +2538,6 @@ class TestToolCallExecution:
             )
 
         callback.assert_not_awaited()
-
-    @pytest.mark.asyncio
-    async def test_sticky_rejection_skips_dynamic_approval_checker(self, mock_model):
-        checker_calls: list[str] = []
-        tool_calls: list[str] = []
-
-        async def needs_approval(_ctx: Any, _params: dict[str, Any], call_id: str) -> bool:
-            checker_calls.append(call_id)
-            if call_id != "call-reject-first":
-                raise AssertionError("sticky rejection must bypass needs_approval")
-            return True
-
-        async def invoke_tool(_ctx: ToolContext[Any], _arguments: str) -> str:
-            tool_calls.append("called")
-            return "should-not-run"
-
-        tool = FunctionTool(
-            name="send_email",
-            description="Send an email.",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=invoke_tool,
-            needs_approval=needs_approval,
-        )
-        agent = RealtimeAgent(name="agent", tools=[tool])
-        session = RealtimeSession(mock_model, agent, None, run_config={"async_tool_calls": False})
-        first_call = RealtimeModelToolCallEvent(
-            name=tool.name, call_id="call-reject-first", arguments="{}"
-        )
-        second_call = RealtimeModelToolCallEvent(
-            name=tool.name, call_id="call-reject-second", arguments="{}"
-        )
-
-        await session._handle_tool_call(first_call)
-        await session.reject_tool_call(first_call.call_id, always=True)
-        await session._handle_tool_call(second_call)
-
-        assert checker_calls == ["call-reject-first"]
-        assert tool_calls == []
-        assert session._pending_tool_calls == {}
-        assert len(mock_model.sent_tool_outputs) == 2
-
-    @pytest.mark.asyncio
-    async def test_sticky_rejection_wins_while_dynamic_approval_checker_is_pending(
-        self, mock_model
-    ):
-        checker_started = asyncio.Event()
-        checker_release = asyncio.Event()
-        checker_calls: list[str] = []
-        tool_calls: list[str] = []
-
-        async def needs_approval(_ctx: Any, _params: dict[str, Any], call_id: str) -> bool:
-            checker_calls.append(call_id)
-            if call_id == "call-pending-checker":
-                checker_started.set()
-                await checker_release.wait()
-                return False
-            return True
-
-        async def invoke_tool(_ctx: ToolContext[Any], _arguments: str) -> str:
-            tool_calls.append("called")
-            return "should-not-run"
-
-        tool = FunctionTool(
-            name="send_email",
-            description="Send an email.",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=invoke_tool,
-            needs_approval=needs_approval,
-        )
-        agent = RealtimeAgent(name="agent", tools=[tool])
-        session = RealtimeSession(mock_model, agent, None)
-        first_call = RealtimeModelToolCallEvent(
-            name=tool.name, call_id="call-reject-first", arguments="{}"
-        )
-        pending_checker_call = RealtimeModelToolCallEvent(
-            name=tool.name, call_id="call-pending-checker", arguments="{}"
-        )
-
-        await session._handle_tool_call(first_call)
-        pending_checker_task = asyncio.create_task(session._handle_tool_call(pending_checker_call))
-        try:
-            await asyncio.wait_for(checker_started.wait(), timeout=1)
-            await session.reject_tool_call(
-                first_call.call_id,
-                always=True,
-                rejection_message="sticky rejection",
-            )
-        finally:
-            checker_release.set()
-        await pending_checker_task
-
-        assert checker_calls == ["call-reject-first", "call-pending-checker"]
-        assert tool_calls == []
-        assert session._pending_tool_calls == {}
-        assert [output for _call, output, _start in mock_model.sent_tool_outputs] == [
-            "sticky rejection",
-            "sticky rejection",
-        ]
-
-    @pytest.mark.parametrize("approved", [True, False], ids=["approved", "rejected"])
-    @pytest.mark.asyncio
-    async def test_sticky_decision_wins_while_rejecting_pre_approval_guardrail_is_pending(
-        self, mock_model, approved: bool
-    ):
-        guardrail_started = asyncio.Event()
-        guardrail_release = asyncio.Event()
-        guardrail_calls: list[str | None] = []
-        tool_calls: list[str] = []
-
-        @tool_input_guardrail
-        async def blocking_guardrail(
-            data: ToolInputGuardrailData,
-        ) -> ToolGuardrailFunctionOutput:
-            call_id = data.context.tool_call_id
-            guardrail_calls.append(call_id)
-            if call_id == "call-pending-guardrail":
-                guardrail_started.set()
-                await guardrail_release.wait()
-                return ToolGuardrailFunctionOutput.reject_content("guardrail rejection")
-            return ToolGuardrailFunctionOutput.allow()
-
-        async def invoke_tool(_ctx: ToolContext[Any], _arguments: str) -> str:
-            tool_calls.append("called")
-            return "tool output"
-
-        tool = FunctionTool(
-            name="send_email",
-            description="Send an email.",
-            params_json_schema={"type": "object", "properties": {}},
-            on_invoke_tool=invoke_tool,
-            needs_approval=True,
-            tool_input_guardrails=[blocking_guardrail],
-        )
-        agent = RealtimeAgent(name="agent", tools=[tool])
-        session = RealtimeSession(
-            mock_model,
-            agent,
-            None,
-            run_config={"tool_execution": {"pre_approval_tool_input_guardrails": True}},
-        )
-        first_call = RealtimeModelToolCallEvent(
-            name=tool.name, call_id="call-reject-first", arguments="{}"
-        )
-        pending_guardrail_call = RealtimeModelToolCallEvent(
-            name=tool.name, call_id="call-pending-guardrail", arguments="{}"
-        )
-
-        await session._handle_tool_call(first_call)
-        pending_guardrail_task = asyncio.create_task(
-            session._handle_tool_call(pending_guardrail_call)
-        )
-        try:
-            await asyncio.wait_for(guardrail_started.wait(), timeout=1)
-            approval_item = session._pending_tool_calls[first_call.call_id].approval_item
-            if approved:
-                session._context_wrapper.approve_tool(approval_item, always_approve=True)
-            else:
-                session._context_wrapper.reject_tool(
-                    approval_item,
-                    always_reject=True,
-                    rejection_message="sticky rejection",
-                )
-        finally:
-            guardrail_release.set()
-        await pending_guardrail_task
-
-        assert pending_guardrail_call.call_id not in session._pending_tool_calls
-        outputs = [output for _call, output, _start in mock_model.sent_tool_outputs]
-        if approved:
-            assert guardrail_calls == [
-                "call-reject-first",
-                "call-pending-guardrail",
-                "call-pending-guardrail",
-            ]
-            assert tool_calls == []
-            assert outputs == ["guardrail rejection"]
-        else:
-            assert guardrail_calls == ["call-reject-first", "call-pending-guardrail"]
-            assert tool_calls == []
-            assert outputs == ["sticky rejection"]
 
     @pytest.mark.asyncio
     async def test_function_tool_exception_handling(
@@ -4591,128 +2622,6 @@ class TestToolCallExecution:
         assert isinstance(tool_context, ToolContext)
 
     @pytest.mark.asyncio
-    async def test_tool_result_conversion_to_string(self, mock_model, mock_agent):
-        """Test that structured tool results are serialized to JSON for model output."""
-        # Create tool that returns non-string result
-        tool = _set_default_timeout_fields(Mock(spec=FunctionTool))
-        tool.name = "test_function"
-        tool.on_invoke_tool = AsyncMock(return_value={"result": "data", "count": 42})
-        tool.needs_approval = False
-
-        mock_agent.get_all_tools.return_value = [tool]
-
-        session = RealtimeSession(mock_model, mock_agent, None)
-
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_conversion", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-
-        # Verify result was serialized to JSON
-        sent_call, sent_output, _ = mock_model.sent_tool_outputs[0]
-        assert isinstance(sent_output, str)
-        assert sent_output == json.dumps({"result": "data", "count": 42})
-
-    @pytest.mark.asyncio
-    async def test_tool_result_conversion_serializes_pydantic_models(self, mock_model, mock_agent):
-        """Test that pydantic tool results are serialized to JSON for model output."""
-
-        class ToolResult(BaseModel):
-            name: str
-            score: int
-
-        tool = _set_default_timeout_fields(Mock(spec=FunctionTool))
-        tool.name = "test_function"
-        tool.on_invoke_tool = AsyncMock(return_value=ToolResult(name="demo", score=7))
-        tool.needs_approval = False
-
-        mock_agent.get_all_tools.return_value = [tool]
-
-        session = RealtimeSession(mock_model, mock_agent, None)
-
-        tool_call_event = RealtimeModelToolCallEvent(
-            name="test_function", call_id="call_pydantic_conversion", arguments="{}"
-        )
-
-        await session._handle_tool_call(tool_call_event)
-
-        _sent_call, sent_output, _ = mock_model.sent_tool_outputs[0]
-        assert sent_output == json.dumps({"name": "demo", "score": 7})
-
-    def test_serialize_tool_output_ignores_non_pydantic_model_dump_objects(self) -> None:
-        class ModelDumpObject:
-            def model_dump(self, *_args: Any, **_kwargs: Any) -> dict[str, Any]:
-                raise AssertionError("non-pydantic objects should not use model_dump")
-
-            def __str__(self) -> str:
-                return "fake-model-dump-object"
-
-        assert _serialize_tool_output(ModelDumpObject()) == "fake-model-dump-object"
-
-    def test_serialize_tool_output_falls_back_when_pydantic_json_dump_fails(self) -> None:
-        class FallbackModel(BaseModel):
-            model_config = ConfigDict(arbitrary_types_allowed=True)
-
-            payload: object
-
-            def model_dump(self, *args: Any, **kwargs: Any) -> dict[str, Any]:
-                if kwargs.get("mode") == "json":
-                    raise ValueError("json mode failed")
-                return {"payload": "ok"}
-
-        assert _serialize_tool_output(FallbackModel(payload=object())) == json.dumps(
-            {"payload": "ok"}
-        )
-
-    def test_serialize_tool_output_returns_string_when_pydantic_dump_fails(self) -> None:
-        class BrokenModel(BaseModel):
-            value: int
-
-            def model_dump(self, *args: Any, **kwargs: Any) -> dict[str, Any]:
-                raise ValueError("dump failed")
-
-            def __str__(self) -> str:
-                return "broken-model"
-
-        assert _serialize_tool_output(BrokenModel(value=1)) == "broken-model"
-
-    def test_serialize_tool_output_returns_string_when_dataclass_asdict_fails(self) -> None:
-        @dataclasses.dataclass
-        class BrokenDataclass:
-            lock: Any
-
-            def __str__(self) -> str:
-                return "broken-dataclass"
-
-        assert _serialize_tool_output(BrokenDataclass(lock=threading.Lock())) == "broken-dataclass"
-
-    @dataclasses.dataclass
-    class ToolResult:
-        label: str
-        values: list[int]
-
-    @pytest.mark.parametrize(
-        ("value", "expected"),
-        [
-            pytest.param(None, "null", id="none"),
-            pytest.param(
-                ["hello", 1, True, None],
-                json.dumps(["hello", 1, True, None]),
-                id="list",
-            ),
-            pytest.param(
-                ToolResult(label="demo", values=[1, 2]),
-                json.dumps({"label": "demo", "values": [1, 2]}),
-                id="dataclass",
-            ),
-            pytest.param(b"abc", "b'abc'", id="bytes"),
-        ],
-    )
-    def test_serialize_tool_output_edge_cases(self, value: Any, expected: str) -> None:
-        assert _serialize_tool_output(value) == expected
-
-    @pytest.mark.asyncio
     async def test_mixed_tool_types_filtering(self, mock_model, mock_agent):
         """Test that function tools and handoffs are properly separated"""
         # Create mixed tools
@@ -4755,1073 +2664,6 @@ class TestToolCallExecution:
         # Verify result
         sent_call, sent_output, _ = mock_model.sent_tool_outputs[0]
         assert sent_output == "result2"
-
-
-class TestGuardrailFunctionality:
-    """Test suite for output guardrail functionality in RealtimeSession"""
-
-    async def _wait_for_guardrail_tasks(self, session):
-        """Wait for all pending guardrail tasks to complete."""
-        import asyncio
-
-        if session._guardrail_tasks:
-            await asyncio.gather(*session._guardrail_tasks, return_exceptions=True)
-
-    @pytest.fixture
-    def triggered_guardrail(self):
-        """Creates a guardrail that always triggers"""
-
-        def guardrail_func(context, agent, output):
-            return GuardrailFunctionOutput(
-                output_info={"reason": "test trigger"}, tripwire_triggered=True
-            )
-
-        return OutputGuardrail(guardrail_function=guardrail_func, name="triggered_guardrail")
-
-    @pytest.fixture
-    def safe_guardrail(self):
-        """Creates a guardrail that never triggers"""
-
-        def guardrail_func(context, agent, output):
-            return GuardrailFunctionOutput(
-                output_info={"reason": "safe content"}, tripwire_triggered=False
-            )
-
-        return OutputGuardrail(guardrail_function=guardrail_func, name="safe_guardrail")
-
-    @pytest.mark.parametrize(
-        ("model_redacted", "tool_redacted"),
-        [(True, False), (False, True), (False, False)],
-        ids=["model_redacted", "tool_redacted", "diagnostic"],
-    )
-    @pytest.mark.asyncio
-    async def test_output_guardrail_failure_follows_both_data_policies(
-        self,
-        monkeypatch: pytest.MonkeyPatch,
-        caplog: pytest.LogCaptureFixture,
-        mock_model: RealtimeModel,
-        model_redacted: bool,
-        tool_redacted: bool,
-    ) -> None:
-        error = RuntimeError("SECRET_REALTIME_GUARDRAIL_ERROR")
-
-        async def failing_guardrail(context, agent, output):
-            _ = context, agent, output
-            raise error
-
-        guardrail = OutputGuardrail(
-            guardrail_function=failing_guardrail,
-            name="SECRET_REALTIME_GUARDRAIL_NAME",
-        )
-        agent = RealtimeAgent(name="agent", output_guardrails=[guardrail])
-        session = RealtimeSession(mock_model, agent, None)
-        monkeypatch.setattr(_debug, "DONT_LOG_MODEL_DATA", model_redacted)
-        monkeypatch.setattr(_debug, "DONT_LOG_TOOL_DATA", tool_redacted)
-
-        with caplog.at_level(logging.DEBUG, logger="openai.agents"):
-            triggered = await session._run_output_guardrails("model text", "response-id")
-
-        assert triggered is False
-        records = [
-            record
-            for record in caplog.records
-            if "Output guardrail raised an exception" in record.getMessage()
-        ]
-        assert len(records) == 1
-        record = records[0]
-        redacted = model_redacted or tool_redacted
-        if redacted:
-            assert record.msg == "%s"
-            assert record.args == ("Output guardrail raised an exception; skipping it",)
-            assert record.exc_info is None
-            assert record.exc_text is None
-            assert "openai_agents_diagnostic_context" not in record.__dict__
-            assert error not in record.__dict__.values()
-            rendered = logging.Formatter().format(record)
-            assert "SECRET_REALTIME_GUARDRAIL_ERROR" not in rendered
-            assert "SECRET_REALTIME_GUARDRAIL_NAME" not in rendered
-        else:
-            context = record.__dict__["openai_agents_diagnostic_context"]
-            assert context == {"guardrail_name": "SECRET_REALTIME_GUARDRAIL_NAME"}
-            assert record.exc_info is not None
-            assert record.exc_info[1] is error
-            assert "SECRET_REALTIME_GUARDRAIL_ERROR" in logging.Formatter().format(record)
-
-    @pytest.mark.asyncio
-    async def test_output_guardrail_failure_tolerates_missing_callable_name(
-        self,
-        monkeypatch: pytest.MonkeyPatch,
-        caplog: pytest.LogCaptureFixture,
-        mock_model: RealtimeModel,
-    ) -> None:
-        class _FailingGuardrailCallable:
-            async def __call__(self, context, agent, output):
-                _ = context, agent, output
-                raise RuntimeError("SECRET_UNNAMED_GUARDRAIL_ERROR")
-
-        guardrail = OutputGuardrail(guardrail_function=_FailingGuardrailCallable())
-        agent = RealtimeAgent(name="agent", output_guardrails=[guardrail])
-        session = RealtimeSession(mock_model, agent, None)
-        monkeypatch.setattr(_debug, "DONT_LOG_MODEL_DATA", False)
-        monkeypatch.setattr(_debug, "DONT_LOG_TOOL_DATA", False)
-
-        with caplog.at_level(logging.WARNING, logger="openai.agents"):
-            triggered = await session._run_output_guardrails("model text", "response-id")
-
-        assert triggered is False
-        records = [
-            record
-            for record in caplog.records
-            if "Output guardrail raised an exception" in record.getMessage()
-        ]
-        assert len(records) == 1
-        context = records[0].__dict__["openai_agents_diagnostic_context"]
-        assert context["guardrail_type"].endswith("._FailingGuardrailCallable")
-        assert records[0].exc_info is not None
-
-    @pytest.mark.asyncio
-    async def test_transcript_delta_triggers_guardrail_at_threshold(
-        self, mock_model, mock_agent, triggered_guardrail
-    ):
-        """Test that guardrails run when transcript delta reaches debounce threshold"""
-        run_config: RealtimeRunConfig = {
-            "output_guardrails": [triggered_guardrail],
-            "guardrails_settings": {"debounce_text_length": 10},
-        }
-
-        session = RealtimeSession(mock_model, mock_agent, None, run_config=run_config)
-
-        # Send transcript delta that exceeds threshold (10 chars)
-        transcript_event = RealtimeModelTranscriptDeltaEvent(
-            item_id="item_1", delta="this is more than ten characters", response_id="resp_1"
-        )
-
-        await session.on_event(transcript_event)
-
-        # Wait for async guardrail tasks to complete
-        await self._wait_for_guardrail_tasks(session)
-
-        # Should have triggered guardrail and interrupted
-        assert mock_model.interrupts_called == 1
-        interrupt_event = next(
-            event
-            for event in mock_model.sent_events
-            if isinstance(event, RealtimeModelSendInterrupt)
-        )
-        assert interrupt_event.force_response_cancel is True
-        assert len(mock_model.sent_messages) == 1
-        assert mock_model.sent_messages[0] == "guardrail triggered: triggered_guardrail"
-
-        # Should have emitted guardrail_tripped event
-        events = []
-        while not session._event_queue.empty():
-            events.append(await session._event_queue.get())
-
-        guardrail_events = [e for e in events if isinstance(e, RealtimeGuardrailTripped)]
-        assert len(guardrail_events) == 1
-        assert guardrail_events[0].message == "this is more than ten characters"
-
-    @pytest.mark.asyncio
-    async def test_output_text_delta_triggers_response_scoped_guardrail(
-        self, mock_model, mock_agent, triggered_guardrail
-    ):
-        run_config: RealtimeRunConfig = {
-            "output_guardrails": [triggered_guardrail],
-            "guardrails_settings": {"debounce_text_length": 5},
-        }
-        session = RealtimeSession(mock_model, mock_agent, None, run_config=run_config)
-
-        await session.on_event(RealtimeModelTurnStartedEvent())
-        await session.on_event(
-            RealtimeModelOutputTextDeltaEvent(
-                item_id="item_1",
-                delta="hello",
-                response_id="response_1",
-            )
-        )
-        await self._wait_for_guardrail_tasks(session)
-
-        interrupt_event = next(
-            event
-            for event in mock_model.sent_events
-            if isinstance(event, RealtimeModelSendInterrupt)
-        )
-        assert interrupt_event.force_response_cancel is True
-        assert interrupt_event.response_id == "response_1"
-        assert interrupt_event.cancel_response_only is True
-        assert mock_model.sent_messages == ["guardrail triggered: triggered_guardrail"]
-
-    @pytest.mark.asyncio
-    async def test_stale_output_text_guardrail_does_not_affect_newer_response(self, mock_model):
-        guardrail_started = asyncio.Event()
-        release_guardrail = asyncio.Event()
-
-        async def delayed_guardrail(context, agent, output):
-            _ = context, agent, output
-            guardrail_started.set()
-            await release_guardrail.wait()
-            return GuardrailFunctionOutput(output_info={}, tripwire_triggered=True)
-
-        guardrail = OutputGuardrail(
-            guardrail_function=delayed_guardrail,
-            name="delayed_guardrail",
-        )
-        source_agent = RealtimeAgent(name="source", output_guardrails=[guardrail])
-        session = RealtimeSession(
-            mock_model,
-            source_agent,
-            None,
-            run_config={"guardrails_settings": {"debounce_text_length": 1}},
-        )
-
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_1"))
-        await session.on_event(
-            RealtimeModelOutputTextDeltaEvent(
-                item_id="item_1",
-                delta="blocked",
-                response_id="response_1",
-            )
-        )
-        await guardrail_started.wait()
-
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_2"))
-        release_guardrail.set()
-        await self._wait_for_guardrail_tasks(session)
-
-        assert not any(
-            isinstance(event, RealtimeModelSendInterrupt) for event in mock_model.sent_events
-        )
-        assert mock_model.sent_messages == []
-        queued_events = []
-        while not session._event_queue.empty():
-            queued_events.append(await session._event_queue.get())
-        assert sum(isinstance(event, RealtimeGuardrailTripped) for event in queued_events) == 1
-
-    @pytest.mark.asyncio
-    async def test_stale_audio_guardrail_interrupts_only_source_playback(self, mock_model):
-        guardrail_started = asyncio.Event()
-        release_guardrail = asyncio.Event()
-
-        async def delayed_guardrail(context, agent, output):
-            _ = context, agent, output
-            guardrail_started.set()
-            await release_guardrail.wait()
-            return GuardrailFunctionOutput(output_info={}, tripwire_triggered=True)
-
-        session = RealtimeSession(
-            mock_model,
-            RealtimeAgent(
-                name="source",
-                output_guardrails=[
-                    OutputGuardrail(
-                        guardrail_function=delayed_guardrail,
-                        name="delayed_guardrail",
-                    )
-                ],
-            ),
-            None,
-            run_config={"guardrails_settings": {"debounce_text_length": 1}},
-        )
-
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_1"))
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(
-                item_id="item_1",
-                delta="blocked",
-                response_id="response_1",
-            )
-        )
-        await guardrail_started.wait()
-        await session.on_event(RealtimeModelTurnEndedEvent(response_id="response_1"))
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_2"))
-
-        assert mock_model.retired_audio_response_ids == []
-        release_guardrail.set()
-        await self._wait_for_guardrail_tasks(session)
-
-        interrupts = [
-            event
-            for event in mock_model.sent_events
-            if isinstance(event, RealtimeModelSendInterrupt)
-        ]
-        assert len(interrupts) == 1
-        assert interrupts[0].response_id == "response_1"
-        assert interrupts[0].playback_only is True
-        assert interrupts[0].force_response_cancel is False
-        assert mock_model.sent_messages == []
-        assert mock_model.retired_audio_response_ids == ["response_1"]
-        assert session._interrupted_response_ids == set()
-
-    @pytest.mark.asyncio
-    async def test_response_audio_cleanup_waits_for_delayed_guardrail(self, mock_agent):
-        guardrail_started = asyncio.Event()
-        release_guardrail = asyncio.Event()
-        operations: list[str] = []
-
-        class TrackingModel(RecordingRealtimeModel):
-            async def send_event(self, event):
-                await super().send_event(event)
-                if isinstance(event, RealtimeModelSendInterrupt):
-                    operations.append("interrupt")
-
-            def _retire_response_audio(self, response_id: str) -> None:
-                super()._retire_response_audio(response_id)
-                operations.append("retire")
-
-        async def delayed_guardrail(context, agent, output):
-            _ = context, agent, output
-            guardrail_started.set()
-            await release_guardrail.wait()
-            return GuardrailFunctionOutput(output_info={}, tripwire_triggered=True)
-
-        model = TrackingModel()
-        session = RealtimeSession(
-            model,
-            mock_agent,
-            None,
-            run_config={
-                "output_guardrails": [
-                    OutputGuardrail(
-                        guardrail_function=delayed_guardrail,
-                        name="delayed_guardrail",
-                    )
-                ],
-                "guardrails_settings": {"debounce_text_length": 1},
-            },
-        )
-
-        await session.on_event(RealtimeModelTurnStartedEvent())
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(
-                item_id="item_1",
-                delta="blocked",
-                response_id="response_1",
-            )
-        )
-        await guardrail_started.wait()
-        assert session._active_output_response_id == "response_1"
-        await session.on_event(RealtimeModelTurnEndedEvent())
-        await asyncio.sleep(0)
-
-        assert operations == []
-        release_guardrail.set()
-        await self._wait_for_guardrail_tasks(session)
-
-        assert operations == ["interrupt", "retire"]
-        assert model.retired_audio_response_ids == ["response_1"]
-        assert session._guardrail_tasks_by_response_id == {}
-        assert session._responses_awaiting_guardrail_cleanup == set()
-
-    @pytest.mark.asyncio
-    async def test_response_audio_cleanup_runs_immediately_without_guardrail_tasks(
-        self, mock_model, mock_agent
-    ):
-        session = RealtimeSession(mock_model, mock_agent, None)
-
-        await session.on_event(RealtimeModelTurnEndedEvent(response_id="response_1"))
-
-        assert mock_model.retired_audio_response_ids == ["response_1"]
-
-    @pytest.mark.asyncio
-    async def test_stale_explicit_turn_end_preserves_active_response_guardrail_state(
-        self, mock_model, mock_agent
-    ):
-        session = RealtimeSession(mock_model, mock_agent, None)
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="new_response"))
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(
-                item_id="new_item",
-                delta="still active",
-                response_id="new_response",
-            )
-        )
-        active_generation = session._active_output_response_generation
-        active_agent = session._active_output_response_agent
-
-        await session.on_event(RealtimeModelTurnEndedEvent(response_id="old_response"))
-
-        assert mock_model.retired_audio_response_ids == ["old_response"]
-        assert session._active_output_response_id == "new_response"
-        assert session._active_output_response_generation == active_generation
-        assert session._active_output_response_agent is active_agent
-        assert session._item_transcripts == {"new_item": "still active"}
-        assert session._item_guardrail_run_counts == {"new_item": 0}
-        queued_events = []
-        while not session._event_queue.empty():
-            queued_events.append(await session._event_queue.get())
-        assert not any(isinstance(event, RealtimeAgentEndEvent) for event in queued_events)
-
-    @pytest.mark.asyncio
-    async def test_interrupted_response_audio_delta_is_not_forwarded(self, mock_model, mock_agent):
-        session = RealtimeSession(mock_model, mock_agent, None)
-        session._interrupted_response_ids.add("response_1")
-
-        await session.on_event(
-            RealtimeModelAudioEvent(
-                data=b"audio",
-                response_id="response_1",
-                item_id="item_1",
-                content_index=0,
-            )
-        )
-
-        queued_events = []
-        while not session._event_queue.empty():
-            queued_events.append(await session._event_queue.get())
-        assert not any(isinstance(event, RealtimeAudio) for event in queued_events)
-
-    @pytest.mark.asyncio
-    async def test_response_audio_cleanup_error_releases_session_suppression(self, mock_agent):
-        class FailingRetirementModel(RecordingRealtimeModel):
-            def _retire_response_audio(self, response_id: str) -> None:
-                raise RuntimeError(f"failed to retire {response_id}")
-
-        session = RealtimeSession(FailingRetirementModel(), mock_agent, None)
-        session._interrupted_response_ids.add("response_1")
-
-        session._retire_response_audio("response_1")
-
-        assert session._interrupted_response_ids == set()
-        queued_event = await session._event_queue.get()
-        assert isinstance(queued_event, RealtimeError)
-        assert queued_event.error == {
-            "message": "Response audio cleanup failed: failed to retire response_1"
-        }
-
-    @pytest.mark.asyncio
-    async def test_output_text_guardrail_sends_feedback_after_source_turn_ends(
-        self, mock_model, mock_agent, triggered_guardrail
-    ):
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={
-                "output_guardrails": [triggered_guardrail],
-                "guardrails_settings": {"debounce_text_length": 5},
-            },
-        )
-        original_send_event = mock_model.send_event
-
-        async def send_event(event):
-            await original_send_event(event)
-            if isinstance(event, RealtimeModelSendInterrupt):
-                await session.on_event(RealtimeModelTurnEndedEvent())
-
-        mock_model.send_event = send_event
-
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_1"))
-        await session.on_event(
-            RealtimeModelOutputTextDeltaEvent(
-                item_id="item_1",
-                delta="hello",
-                response_id="response_1",
-            )
-        )
-        await self._wait_for_guardrail_tasks(session)
-
-        assert mock_model.sent_messages == ["guardrail triggered: triggered_guardrail"]
-
-    @pytest.mark.asyncio
-    async def test_output_text_guardrail_skips_feedback_for_completed_idless_newer_turn(
-        self, mock_model, mock_agent, triggered_guardrail
-    ):
-        session = RealtimeSession(
-            mock_model,
-            mock_agent,
-            None,
-            run_config={
-                "output_guardrails": [triggered_guardrail],
-                "guardrails_settings": {"debounce_text_length": 5},
-            },
-        )
-        original_send_event = mock_model.send_event
-
-        async def send_event(event):
-            await original_send_event(event)
-            if isinstance(event, RealtimeModelSendInterrupt):
-                await session.on_event(RealtimeModelTurnEndedEvent())
-                await session.on_event(RealtimeModelTurnStartedEvent())
-                await session.on_event(RealtimeModelTurnEndedEvent())
-
-        mock_model.send_event = send_event
-
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_1"))
-        await session.on_event(
-            RealtimeModelOutputTextDeltaEvent(
-                item_id="item_1",
-                delta="hello",
-                response_id="response_1",
-            )
-        )
-        await self._wait_for_guardrail_tasks(session)
-
-        assert mock_model.sent_messages == []
-
-    @pytest.mark.asyncio
-    async def test_output_text_guardrail_rechecks_generation_at_feedback_send_boundary(
-        self, mock_agent, triggered_guardrail
-    ):
-        feedback_send_started = asyncio.Event()
-        release_feedback_send = asyncio.Event()
-
-        class BoundaryCheckingModel(RecordingRealtimeModel):
-            async def send_event_if(self, event, send_if):
-                feedback_send_started.set()
-                await release_feedback_send.wait()
-                return await super().send_event_if(event, send_if)
-
-        model = BoundaryCheckingModel()
-        session = RealtimeSession(
-            model,
-            mock_agent,
-            None,
-            run_config={
-                "output_guardrails": [triggered_guardrail],
-                "guardrails_settings": {"debounce_text_length": 5},
-            },
-        )
-
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_1"))
-        await session.on_event(
-            RealtimeModelOutputTextDeltaEvent(
-                item_id="item_1",
-                delta="hello",
-                response_id="response_1",
-            )
-        )
-        await feedback_send_started.wait()
-
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_2"))
-        release_feedback_send.set()
-        await self._wait_for_guardrail_tasks(session)
-
-        assert model.sent_messages == []
-
-    @pytest.mark.asyncio
-    async def test_output_text_guardrail_skips_feedback_without_atomic_model_send(
-        self, mock_agent, triggered_guardrail
-    ):
-        class CustomModelWithoutAtomicSend(RecordingRealtimeModel):
-            def __init__(self):
-                super().__init__()
-                self.feedback_send_started = False
-
-            async def send_event(self, event):
-                if isinstance(event, RealtimeModelSendUserInput):
-                    self.feedback_send_started = True
-                    await asyncio.sleep(0)
-                await super().send_event(event)
-
-            async def send_event_if(self, event, send_if):
-                return await RealtimeModel.send_event_if(self, event, send_if)
-
-        model = CustomModelWithoutAtomicSend()
-        session = RealtimeSession(
-            model,
-            mock_agent,
-            None,
-            run_config={
-                "output_guardrails": [triggered_guardrail],
-                "guardrails_settings": {"debounce_text_length": 5},
-            },
-        )
-
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_1"))
-        await session.on_event(
-            RealtimeModelOutputTextDeltaEvent(
-                item_id="item_1",
-                delta="hello",
-                response_id="response_1",
-            )
-        )
-        await self._wait_for_guardrail_tasks(session)
-
-        assert any(isinstance(event, RealtimeModelSendInterrupt) for event in model.sent_events)
-        assert model.feedback_send_started is False
-        assert model.sent_messages == []
-
-    @pytest.mark.asyncio
-    async def test_output_text_guardrail_uses_agent_from_turn_start(self, mock_model):
-        observed_agents: list[RealtimeAgent] = []
-        replacement_called = False
-
-        def source_guardrail(context, agent, output):
-            _ = context, output
-            observed_agents.append(agent)
-            return GuardrailFunctionOutput(output_info={}, tripwire_triggered=True)
-
-        def replacement_guardrail(context, agent, output):
-            nonlocal replacement_called
-            _ = context, agent, output
-            replacement_called = True
-            return GuardrailFunctionOutput(output_info={}, tripwire_triggered=False)
-
-        source_agent = RealtimeAgent(
-            name="source",
-            output_guardrails=[
-                OutputGuardrail(guardrail_function=source_guardrail, name="source_guardrail")
-            ],
-        )
-        replacement_agent = RealtimeAgent(
-            name="replacement",
-            output_guardrails=[
-                OutputGuardrail(
-                    guardrail_function=replacement_guardrail,
-                    name="replacement_guardrail",
-                )
-            ],
-        )
-        session = RealtimeSession(
-            mock_model,
-            source_agent,
-            None,
-            run_config={"guardrails_settings": {"debounce_text_length": 5}},
-        )
-
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_1"))
-        await session.update_agent(replacement_agent)
-        await session.on_event(
-            RealtimeModelOutputTextDeltaEvent(
-                item_id="item_1",
-                delta="hello",
-                response_id="response_1",
-            )
-        )
-        await self._wait_for_guardrail_tasks(session)
-
-        assert observed_agents == [source_agent]
-        assert replacement_called is False
-        assert mock_model.sent_messages == ["guardrail triggered: source_guardrail"]
-
-    @pytest.mark.asyncio
-    async def test_output_text_guardrail_retains_agent_for_matching_late_turn_start(
-        self, mock_model
-    ):
-        observed_agents: list[RealtimeAgent] = []
-
-        def source_guardrail(context, agent, output):
-            _ = context, output
-            observed_agents.append(agent)
-            return GuardrailFunctionOutput(output_info={}, tripwire_triggered=True)
-
-        source_agent = RealtimeAgent(
-            name="source",
-            output_guardrails=[
-                OutputGuardrail(guardrail_function=source_guardrail, name="source_guardrail")
-            ],
-        )
-        replacement_agent = RealtimeAgent(name="replacement")
-        session = RealtimeSession(
-            mock_model,
-            source_agent,
-            None,
-            run_config={"guardrails_settings": {"debounce_text_length": 5}},
-        )
-
-        await session.on_event(
-            RealtimeModelOutputTextDeltaEvent(
-                item_id="item_1",
-                delta="he",
-                response_id="response_1",
-            )
-        )
-        await session.update_agent(replacement_agent)
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_1"))
-        await session.on_event(
-            RealtimeModelOutputTextDeltaEvent(
-                item_id="item_1",
-                delta="llo",
-                response_id="response_1",
-            )
-        )
-        await self._wait_for_guardrail_tasks(session)
-
-        assert observed_agents == [source_agent]
-        assert mock_model.sent_messages == ["guardrail triggered: source_guardrail"]
-
-    @pytest.mark.asyncio
-    async def test_matching_late_turn_start_retains_pending_guardrail_generation(self, mock_model):
-        guardrail_started = asyncio.Event()
-        release_guardrail = asyncio.Event()
-
-        async def delayed_guardrail(context, agent, output):
-            _ = context, agent, output
-            guardrail_started.set()
-            await release_guardrail.wait()
-            return GuardrailFunctionOutput(output_info={}, tripwire_triggered=True)
-
-        source_agent = RealtimeAgent(
-            name="source",
-            output_guardrails=[
-                OutputGuardrail(guardrail_function=delayed_guardrail, name="source_guardrail")
-            ],
-        )
-        session = RealtimeSession(
-            mock_model,
-            source_agent,
-            None,
-            run_config={"guardrails_settings": {"debounce_text_length": 2}},
-        )
-
-        await session.on_event(
-            RealtimeModelOutputTextDeltaEvent(
-                item_id="item_1",
-                delta="he",
-                response_id="response_1",
-            )
-        )
-        await guardrail_started.wait()
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_1"))
-
-        release_guardrail.set()
-        await self._wait_for_guardrail_tasks(session)
-
-        interrupt_event = next(
-            event
-            for event in mock_model.sent_events
-            if isinstance(event, RealtimeModelSendInterrupt)
-        )
-        assert interrupt_event.response_id == "response_1"
-        assert mock_model.sent_messages == ["guardrail triggered: source_guardrail"]
-
-    @pytest.mark.asyncio
-    async def test_output_text_guardrail_sends_feedback_if_source_ends_during_evaluation(
-        self, mock_model
-    ):
-        guardrail_started = asyncio.Event()
-        release_guardrail = asyncio.Event()
-
-        async def delayed_guardrail(context, agent, output):
-            _ = context, agent, output
-            guardrail_started.set()
-            await release_guardrail.wait()
-            return GuardrailFunctionOutput(output_info={}, tripwire_triggered=True)
-
-        guardrail = OutputGuardrail(
-            guardrail_function=delayed_guardrail,
-            name="delayed_guardrail",
-        )
-        session = RealtimeSession(
-            mock_model,
-            RealtimeAgent(name="source", output_guardrails=[guardrail]),
-            None,
-            run_config={"guardrails_settings": {"debounce_text_length": 1}},
-        )
-
-        await session.on_event(RealtimeModelTurnStartedEvent(response_id="response_1"))
-        await session.on_event(
-            RealtimeModelOutputTextDeltaEvent(
-                item_id="item_1",
-                delta="blocked",
-                response_id="response_1",
-            )
-        )
-        await guardrail_started.wait()
-
-        await session.on_event(RealtimeModelTurnEndedEvent())
-        release_guardrail.set()
-        await self._wait_for_guardrail_tasks(session)
-
-        assert not any(
-            isinstance(event, RealtimeModelSendInterrupt) for event in mock_model.sent_events
-        )
-        assert mock_model.sent_messages == ["guardrail triggered: delayed_guardrail"]
-
-    @pytest.mark.asyncio
-    async def test_agent_and_run_config_guardrails_not_run_twice(self, mock_model):
-        """Guardrails shared by agent and run config should execute once."""
-
-        call_count = 0
-
-        def guardrail_func(context, agent, output):
-            nonlocal call_count
-            call_count += 1
-            return GuardrailFunctionOutput(output_info={}, tripwire_triggered=False)
-
-        shared_guardrail = OutputGuardrail(
-            guardrail_function=guardrail_func, name="shared_guardrail"
-        )
-
-        agent = RealtimeAgent(name="agent", output_guardrails=[shared_guardrail])
-        run_config: RealtimeRunConfig = {
-            "output_guardrails": [shared_guardrail],
-            "guardrails_settings": {"debounce_text_length": 5},
-        }
-
-        session = RealtimeSession(mock_model, agent, None, run_config=run_config)
-
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(item_id="item_1", delta="hello", response_id="resp_1")
-        )
-
-        await self._wait_for_guardrail_tasks(session)
-
-        assert call_count == 1
-
-    @pytest.mark.asyncio
-    async def test_transcript_delta_multiple_thresholds_same_item(
-        self, mock_model, mock_agent, triggered_guardrail
-    ):
-        """Test guardrails run at 1x, 2x, 3x thresholds for same item_id"""
-        run_config: RealtimeRunConfig = {
-            "output_guardrails": [triggered_guardrail],
-            "guardrails_settings": {"debounce_text_length": 5},
-        }
-
-        session = RealtimeSession(mock_model, mock_agent, None, run_config=run_config)
-
-        # First delta - reaches 1x threshold (5 chars)
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(item_id="item_1", delta="12345", response_id="resp_1")
-        )
-
-        # Second delta - reaches 2x threshold (10 chars total)
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(item_id="item_1", delta="67890", response_id="resp_1")
-        )
-
-        # Wait for async guardrail tasks to complete
-        await self._wait_for_guardrail_tasks(session)
-
-        # Should only trigger once due to interrupted_by_guardrail flag
-        assert mock_model.interrupts_called == 1
-        assert len(mock_model.sent_messages) == 1
-
-    @pytest.mark.asyncio
-    async def test_large_transcript_delta_advances_past_each_crossed_threshold(
-        self, mock_model, mock_agent
-    ):
-        calls = 0
-
-        async def guardrail_func(context, agent, output):
-            nonlocal calls
-            calls += 1
-            return GuardrailFunctionOutput(output_info={}, tripwire_triggered=False)
-
-        guardrail = OutputGuardrail(guardrail_function=guardrail_func)
-        run_config: RealtimeRunConfig = {
-            "output_guardrails": [guardrail],
-            "guardrails_settings": {"debounce_text_length": 5},
-        }
-        session = RealtimeSession(mock_model, mock_agent, None, run_config=run_config)
-
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(
-                item_id="item_1", delta="123456789012", response_id="resp_1"
-            )
-        )
-        await self._wait_for_guardrail_tasks(session)
-        assert calls == 1
-
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(item_id="item_1", delta="3", response_id="resp_1")
-        )
-        await self._wait_for_guardrail_tasks(session)
-
-        assert calls == 1
-
-    @pytest.mark.asyncio
-    async def test_transcript_delta_different_items_tracked_separately(
-        self, mock_model, mock_agent, safe_guardrail
-    ):
-        """Test that different item_ids are tracked separately for debouncing"""
-        run_config: RealtimeRunConfig = {
-            "output_guardrails": [safe_guardrail],
-            "guardrails_settings": {"debounce_text_length": 10},
-        }
-
-        session = RealtimeSession(mock_model, mock_agent, None, run_config=run_config)
-
-        # Add text to item_1 (8 chars - below threshold)
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(
-                item_id="item_1", delta="12345678", response_id="resp_1"
-            )
-        )
-
-        # Add text to item_2 (8 chars - below threshold)
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(
-                item_id="item_2", delta="abcdefgh", response_id="resp_2"
-            )
-        )
-
-        # Neither should trigger guardrails yet
-        assert mock_model.interrupts_called == 0
-
-        # Add more text to item_1 (total 12 chars - above threshold)
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(item_id="item_1", delta="90ab", response_id="resp_1")
-        )
-
-        # item_1 should have triggered guardrail run (but not interrupted since safe)
-        assert session._item_guardrail_run_counts["item_1"] == 1
-        assert (
-            "item_2" not in session._item_guardrail_run_counts
-            or session._item_guardrail_run_counts["item_2"] == 0
-        )
-
-    @pytest.mark.asyncio
-    async def test_turn_ended_clears_guardrail_state(
-        self, mock_model, mock_agent, triggered_guardrail
-    ):
-        """Test that turn_ended event clears guardrail state for next turn"""
-        run_config: RealtimeRunConfig = {
-            "output_guardrails": [triggered_guardrail],
-            "guardrails_settings": {"debounce_text_length": 5},
-        }
-
-        session = RealtimeSession(mock_model, mock_agent, None, run_config=run_config)
-
-        # Trigger guardrail
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(
-                item_id="item_1", delta="trigger", response_id="resp_1"
-            )
-        )
-
-        # Wait for async guardrail tasks to complete
-        await self._wait_for_guardrail_tasks(session)
-
-        assert len(session._item_transcripts) == 1
-
-        # End turn
-        await session.on_event(RealtimeModelTurnEndedEvent())
-
-        # State should be cleared
-        assert len(session._item_transcripts) == 0
-        assert len(session._item_guardrail_run_counts) == 0
-
-    @pytest.mark.asyncio
-    async def test_multiple_guardrails_all_triggered(self, mock_model, mock_agent):
-        """Test that all triggered guardrails are included in the event"""
-
-        def create_triggered_guardrail(name):
-            def guardrail_func(context, agent, output):
-                return GuardrailFunctionOutput(output_info={"name": name}, tripwire_triggered=True)
-
-            return OutputGuardrail(guardrail_function=guardrail_func, name=name)
-
-        guardrail1 = create_triggered_guardrail("guardrail_1")
-        guardrail2 = create_triggered_guardrail("guardrail_2")
-
-        run_config: RealtimeRunConfig = {
-            "output_guardrails": [guardrail1, guardrail2],
-            "guardrails_settings": {"debounce_text_length": 5},
-        }
-
-        session = RealtimeSession(mock_model, mock_agent, None, run_config=run_config)
-
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(
-                item_id="item_1", delta="trigger", response_id="resp_1"
-            )
-        )
-
-        # Wait for async guardrail tasks to complete
-        await self._wait_for_guardrail_tasks(session)
-
-        # Should have interrupted and sent message with both guardrail names
-        assert mock_model.interrupts_called == 1
-        assert len(mock_model.sent_messages) == 1
-        message = mock_model.sent_messages[0]
-        assert "guardrail_1" in message and "guardrail_2" in message
-
-        # Should have emitted event with both guardrail results
-        events = []
-        while not session._event_queue.empty():
-            events.append(await session._event_queue.get())
-
-        guardrail_events = [e for e in events if isinstance(e, RealtimeGuardrailTripped)]
-        assert len(guardrail_events) == 1
-        assert len(guardrail_events[0].guardrail_results) == 2
-
-    @pytest.mark.asyncio
-    async def test_agent_output_guardrails_triggered(self, mock_model, triggered_guardrail):
-        """Test that guardrails defined on the agent are executed."""
-        agent = RealtimeAgent(name="agent", output_guardrails=[triggered_guardrail])
-        run_config: RealtimeRunConfig = {
-            "guardrails_settings": {"debounce_text_length": 10},
-        }
-
-        session = RealtimeSession(mock_model, agent, None, run_config=run_config)
-
-        transcript_event = RealtimeModelTranscriptDeltaEvent(
-            item_id="item_1", delta="this is more than ten characters", response_id="resp_1"
-        )
-
-        await session.on_event(transcript_event)
-        await self._wait_for_guardrail_tasks(session)
-
-        assert mock_model.interrupts_called == 1
-        assert len(mock_model.sent_messages) == 1
-        assert "triggered_guardrail" in mock_model.sent_messages[0]
-
-        events = []
-        while not session._event_queue.empty():
-            events.append(await session._event_queue.get())
-
-        guardrail_events = [e for e in events if isinstance(e, RealtimeGuardrailTripped)]
-        assert len(guardrail_events) == 1
-        assert guardrail_events[0].message == "this is more than ten characters"
-
-    @pytest.mark.asyncio
-    async def test_concurrent_guardrail_tasks_interrupt_once_per_response(self, mock_model):
-        """Even if multiple guardrail tasks trigger concurrently for the same response_id,
-        only the first should interrupt and send a message."""
-        import asyncio
-
-        # Barrier to release both guardrail tasks at the same time
-        start_event = asyncio.Event()
-
-        async def async_trigger_guardrail(context, agent, output):
-            await start_event.wait()
-            return GuardrailFunctionOutput(
-                output_info={"reason": "concurrent"}, tripwire_triggered=True
-            )
-
-        concurrent_guardrail = OutputGuardrail(
-            guardrail_function=async_trigger_guardrail, name="concurrent_trigger"
-        )
-
-        run_config: RealtimeRunConfig = {
-            "output_guardrails": [concurrent_guardrail],
-            "guardrails_settings": {"debounce_text_length": 5},
-        }
-
-        # Use a minimal agent (guardrails from run_config)
-        agent = RealtimeAgent(name="agent")
-        session = RealtimeSession(mock_model, agent, None, run_config=run_config)
-
-        # Two deltas for same item and response to enqueue two guardrail tasks
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(
-                item_id="item_1", delta="12345", response_id="resp_same"
-            )
-        )
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(
-                item_id="item_1", delta="67890", response_id="resp_same"
-            )
-        )
-
-        # Wait until both tasks are enqueued
-        for _ in range(50):
-            if len(session._guardrail_tasks) >= 2:
-                break
-            await asyncio.sleep(0.01)
-
-        # Release both tasks concurrently
-        start_event.set()
-
-        # Wait for completion
-        if session._guardrail_tasks:
-            await asyncio.gather(*session._guardrail_tasks, return_exceptions=True)
-
-        # Only one interrupt and one message should be sent
-        assert mock_model.interrupts_called == 1
-        assert len(mock_model.sent_messages) == 1
 
 
 class TestModelSettingsIntegration:
@@ -6176,103 +3018,3 @@ class TestUpdateAgentFunctionality:
 
         assert session._current_agent is first_agent
         assert mock_model.sent_events == ()
-
-
-class TestTranscriptPreservation:
-    """Tests ensuring assistant transcripts are preserved across updates."""
-
-    @pytest.mark.asyncio
-    async def test_assistant_transcript_preserved_on_item_update(self, mock_model, mock_agent):
-        session = RealtimeSession(mock_model, mock_agent, None)
-
-        # Initial assistant message with audio transcript present (e.g., from first turn)
-        initial_item = AssistantMessageItem(
-            item_id="assist_1",
-            role="assistant",
-            content=[AssistantAudio(audio=None, transcript="Hello there")],
-        )
-        session._history = [initial_item]
-
-        # Later, the platform retrieves/updates the same item but without transcript populated
-        updated_without_transcript = AssistantMessageItem(
-            item_id="assist_1",
-            role="assistant",
-            content=[AssistantAudio(audio=None, transcript=None)],
-        )
-
-        await session.on_event(RealtimeModelItemUpdatedEvent(item=updated_without_transcript))
-
-        # Transcript should be preserved from existing history
-        assert len(session._history) == 1
-        preserved_item = cast(AssistantMessageItem, session._history[0])
-        assert isinstance(preserved_item.content[0], AssistantAudio)
-        assert preserved_item.content[0].transcript == "Hello there"
-
-    @pytest.mark.asyncio
-    async def test_assistant_transcript_can_fallback_to_deltas(self, mock_model, mock_agent):
-        session = RealtimeSession(mock_model, mock_agent, None)
-
-        # Simulate transcript deltas accumulated for an assistant item during generation
-        await session.on_event(
-            RealtimeModelTranscriptDeltaEvent(
-                item_id="assist_2", delta="partial transcript", response_id="resp_2"
-            )
-        )
-
-        # Add initial assistant message without transcript
-        initial_item = AssistantMessageItem(
-            item_id="assist_2",
-            role="assistant",
-            content=[AssistantAudio(audio=None, transcript=None)],
-        )
-        await session.on_event(RealtimeModelItemUpdatedEvent(item=initial_item))
-
-        # Later update still lacks transcript; merge should fallback to accumulated deltas
-        update_again = AssistantMessageItem(
-            item_id="assist_2",
-            role="assistant",
-            content=[AssistantAudio(audio=None, transcript=None)],
-        )
-        await session.on_event(RealtimeModelItemUpdatedEvent(item=update_again))
-
-        preserved_item = cast(AssistantMessageItem, session._history[0])
-        assert isinstance(preserved_item.content[0], AssistantAudio)
-        assert preserved_item.content[0].transcript == "partial transcript"
-
-    @pytest.mark.asyncio
-    async def test_existing_transcript_not_overwritten_by_stale_deltas(
-        self, mock_model, mock_agent
-    ):
-        """Existing transcripts must take precedence over leftover delta accumulators.
-
-        ``_item_transcripts`` is keyed by item_id and persists across updates within a
-        turn. When the model retrieves an item without a transcript, the merge should
-        fall back to deltas only when no existing transcript is present – otherwise
-        the complete transcript already in history would be clobbered by partial
-        (or stale) delta state.
-        """
-        session = RealtimeSession(mock_model, mock_agent, None)
-
-        # History already has the completed transcript for the item.
-        initial_item = AssistantMessageItem(
-            item_id="assist_3",
-            role="assistant",
-            content=[AssistantAudio(audio=None, transcript="Final complete transcript")],
-        )
-        session._history = [initial_item]
-
-        # Simulate stale/leftover delta state for the same item id.
-        session._item_transcripts["assist_3"] = "stale partial"
-
-        # Update arrives without transcript populated; merge must keep the existing
-        # complete transcript rather than reverting to the stale delta accumulator.
-        update_without_transcript = AssistantMessageItem(
-            item_id="assist_3",
-            role="assistant",
-            content=[AssistantAudio(audio=None, transcript=None)],
-        )
-        await session.on_event(RealtimeModelItemUpdatedEvent(item=update_without_transcript))
-
-        preserved_item = cast(AssistantMessageItem, session._history[0])
-        assert isinstance(preserved_item.content[0], AssistantAudio)
-        assert preserved_item.content[0].transcript == "Final complete transcript"
